Extract IHTML interfaces to Interop/Mshtml
