+++ src/Common/src/Interop/Mshtml/Interop.IHTMLDocument2.cs
@@ -1,0 +1,129 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Runtime.InteropServices;
+
+internal partial class Interop
+{
+    internal partial class Mshtml
+    {
+        [ComVisible(true)]
+        [Guid("332C4425-26CB-11D0-B483-00C04FD90119")]
+        [InterfaceType(ComInterfaceType.InterfaceIsDual)]
+        internal interface IHTMLDocument2
+        {
+            [return: MarshalAs(UnmanagedType.IDispatch)] object GetScript();
+            IHTMLElementCollection GetAll();
+            IHTMLElement GetBody();
+            IHTMLElement GetActiveElement();
+            IHTMLElementCollection GetImages();
+            IHTMLElementCollection GetApplets();
+            IHTMLElementCollection GetLinks();
+            IHTMLElementCollection GetForms();
+            IHTMLElementCollection GetAnchors();
+            void SetTitle(string p);
+            string GetTitle();
+            IHTMLElementCollection GetScripts();
+            void SetDesignMode(string p);
+            string GetDesignMode();
+            [return: MarshalAs(UnmanagedType.Interface)] object GetSelection();
+            string GetReadyState();
+            [return: MarshalAs(UnmanagedType.Interface)] object GetFrames();
+            IHTMLElementCollection GetEmbeds();
+            IHTMLElementCollection GetPlugins();
+            void SetAlinkColor(object c);
+            object GetAlinkColor();
+            void SetBgColor(object c);
+            object GetBgColor();
+            void SetFgColor(object c);
+            object GetFgColor();
+            void SetLinkColor(object c);
+            object GetLinkColor();
+            void SetVlinkColor(object c);
+            object GetVlinkColor();
+            string GetReferrer();
+            [return: MarshalAs(UnmanagedType.Interface)] IHTMLLocation GetLocation();
+            string GetLastModified();
+            void SetUrl(string p);
+            string GetUrl();
+            void SetDomain(string p);
+            string GetDomain();
+            void SetCookie(string p);
+            string GetCookie();
+            void SetExpando(bool p);
+            bool GetExpando();
+            void SetCharset(string p);
+            string GetCharset();
+            void SetDefaultCharset(string p);
+            string GetDefaultCharset();
+            string GetMimeType();
+            string GetFileSize();
+            string GetFileCreatedDate();
+            string GetFileModifiedDate();
+            string GetFileUpdatedDate();
+            string GetSecurity();
+            string GetProtocol();
+            string GetNameProp();
+            int Write([In, MarshalAs(UnmanagedType.SafeArray)] object[] psarray);
+            int WriteLine([In, MarshalAs(UnmanagedType.SafeArray)] object[] psarray);
+            [return: MarshalAs(UnmanagedType.Interface)] object Open(string mimeExtension, object name, object features, object replace);
+            void Close();
+            void Clear();
+            bool QueryCommandSupported(string cmdID);
+            bool QueryCommandEnabled(string cmdID);
+            bool QueryCommandState(string cmdID);
+            bool QueryCommandIndeterm(string cmdID);
+            string QueryCommandText(
+                    string cmdID);
+            object QueryCommandValue(string cmdID);
+            bool ExecCommand(string cmdID,
+                    bool showUI, object value);
+            bool ExecCommandShowHelp(string cmdID);
+            IHTMLElement CreateElement(string eTag);
+            void SetOnhelp(object p);
+            object GetOnhelp();
+            void SetOnclick(object p);
+            object GetOnclick();
+            void SetOndblclick(object p);
+            object GetOndblclick();
+            void SetOnkeyup(object p);
+            object GetOnkeyup();
+            void SetOnkeydown(object p);
+            object GetOnkeydown();
+            void SetOnkeypress(object p);
+            object GetOnkeypress();
+            void SetOnmouseup(object p);
+            object GetOnmouseup();
+            void SetOnmousedown(object p);
+            object GetOnmousedown();
+            void SetOnmousemove(object p);
+            object GetOnmousemove();
+            void SetOnmouseout(object p);
+            object GetOnmouseout();
+            void SetOnmouseover(object p);
+            object GetOnmouseover();
+            void SetOnreadystatechange(object p);
+            object GetOnreadystatechange();
+            void SetOnafterupdate(object p);
+            object GetOnafterupdate();
+            void SetOnrowexit(object p);
+            object GetOnrowexit();
+            void SetOnrowenter(object p);
+            object GetOnrowenter();
+            void SetOndragstart(object p);
+            object GetOndragstart();
+            void SetOnselectstart(object p);
+            object GetOnselectstart();
+            IHTMLElement ElementFromPoint(int x, int y);
+            [return: MarshalAs(UnmanagedType.Interface)] IHTMLWindow2 GetParentWindow();
+            [return: MarshalAs(UnmanagedType.Interface)] object GetStyleSheets();
+            void SetOnbeforeupdate(object p);
+            object GetOnbeforeupdate();
+            void SetOnerrorupdate(object p);
+            object GetOnerrorupdate();
+            string toString();
+            [return: MarshalAs(UnmanagedType.Interface)] object CreateStyleSheet(string bstrHref, int lIndex);
+        }
+    }
+}
