+++ src/Common/src/Interop/Mshtml/Interop.IHTMLScreen.cs
@@ -1,0 +1,28 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Runtime.InteropServices;
+
+internal partial class Interop
+{
+    internal partial class Mshtml
+    {
+        [ComVisible(true)]
+        [Guid("3050f35c-98b5-11cf-bb82-00aa00bdce0b")]
+        [InterfaceType(ComInterfaceType.InterfaceIsDual)]
+        public interface IHTMLScreen
+        {
+            int GetColorDepth();
+            void SetBufferDepth(int d);
+            int GetBufferDepth();
+            int GetWidth();
+            int GetHeight();
+            void SetUpdateInterval(int i);
+            int GetUpdateInterval();
+            int GetAvailHeight();
+            int GetAvailWidth();
+            bool GetFontSmoothingEnabled();
+        }
+    }
+}
