+++ src/Common/src/UnsafeNativeMethods.cs
@@ -1776,2434 +1776,6 @@ namespace System.Windows.Forms
             [DispId(272)] void PrivacyImpactedStateChange([In] bool bImpacted);
         }
 
-        [ComVisible(true), Guid("626FC520-A41E-11cf-A731-00A0C9082637"),
-        InterfaceType(ComInterfaceType.InterfaceIsDual)]
-        internal interface IHTMLDocument
-        {
-            [return: MarshalAs(UnmanagedType.IDispatch)] object GetScript();
-        }
-
-        [ComVisible(true), Guid("332C4425-26CB-11D0-B483-00C04FD90119"),
-        InterfaceType(ComInterfaceType.InterfaceIsDual)]
-        internal interface IHTMLDocument2
-        {
-            [return: MarshalAs(UnmanagedType.IDispatch)] object GetScript();
-            IHTMLElementCollection GetAll();
-            IHTMLElement GetBody();
-            IHTMLElement GetActiveElement();
-            IHTMLElementCollection GetImages();
-            IHTMLElementCollection GetApplets();
-            IHTMLElementCollection GetLinks();
-            IHTMLElementCollection GetForms();
-            IHTMLElementCollection GetAnchors();
-            void SetTitle(string p);
-            string GetTitle();
-            IHTMLElementCollection GetScripts();
-            void SetDesignMode(string p);
-            string GetDesignMode();
-            [return: MarshalAs(UnmanagedType.Interface)] object GetSelection();
-            string GetReadyState();
-            [return: MarshalAs(UnmanagedType.Interface)] object GetFrames();
-            IHTMLElementCollection GetEmbeds();
-            IHTMLElementCollection GetPlugins();
-            void SetAlinkColor(object c);
-            object GetAlinkColor();
-            void SetBgColor(object c);
-            object GetBgColor();
-            void SetFgColor(object c);
-            object GetFgColor();
-            void SetLinkColor(object c);
-            object GetLinkColor();
-            void SetVlinkColor(object c);
-            object GetVlinkColor();
-            string GetReferrer();
-            [return: MarshalAs(UnmanagedType.Interface)] IHTMLLocation GetLocation();
-            string GetLastModified();
-            void SetUrl(string p);
-            string GetUrl();
-            void SetDomain(string p);
-            string GetDomain();
-            void SetCookie(string p);
-            string GetCookie();
-            void SetExpando(bool p);
-            bool GetExpando();
-            void SetCharset(string p);
-            string GetCharset();
-            void SetDefaultCharset(string p);
-            string GetDefaultCharset();
-            string GetMimeType();
-            string GetFileSize();
-            string GetFileCreatedDate();
-            string GetFileModifiedDate();
-            string GetFileUpdatedDate();
-            string GetSecurity();
-            string GetProtocol();
-            string GetNameProp();
-            int Write([In, MarshalAs(UnmanagedType.SafeArray)] object[] psarray);
-            int WriteLine([In, MarshalAs(UnmanagedType.SafeArray)] object[] psarray);
-            [return: MarshalAs(UnmanagedType.Interface)] object Open(string mimeExtension, object name, object features, object replace);
-            void Close();
-            void Clear();
-            bool QueryCommandSupported(string cmdID);
-            bool QueryCommandEnabled(string cmdID);
-            bool QueryCommandState(string cmdID);
-            bool QueryCommandIndeterm(string cmdID);
-            string QueryCommandText(
-                    string cmdID);
-            object QueryCommandValue(string cmdID);
-            bool ExecCommand(string cmdID,
-                    bool showUI, object value);
-            bool ExecCommandShowHelp(string cmdID);
-            IHTMLElement CreateElement(string eTag);
-            void SetOnhelp(object p);
-            object GetOnhelp();
-            void SetOnclick(object p);
-            object GetOnclick();
-            void SetOndblclick(object p);
-            object GetOndblclick();
-            void SetOnkeyup(object p);
-            object GetOnkeyup();
-            void SetOnkeydown(object p);
-            object GetOnkeydown();
-            void SetOnkeypress(object p);
-            object GetOnkeypress();
-            void SetOnmouseup(object p);
-            object GetOnmouseup();
-            void SetOnmousedown(object p);
-            object GetOnmousedown();
-            void SetOnmousemove(object p);
-            object GetOnmousemove();
-            void SetOnmouseout(object p);
-            object GetOnmouseout();
-            void SetOnmouseover(object p);
-            object GetOnmouseover();
-            void SetOnreadystatechange(object p);
-            object GetOnreadystatechange();
-            void SetOnafterupdate(object p);
-            object GetOnafterupdate();
-            void SetOnrowexit(object p);
-            object GetOnrowexit();
-            void SetOnrowenter(object p);
-            object GetOnrowenter();
-            void SetOndragstart(object p);
-            object GetOndragstart();
-            void SetOnselectstart(object p);
-            object GetOnselectstart();
-            IHTMLElement ElementFromPoint(int x, int y);
-            [return: MarshalAs(UnmanagedType.Interface)] IHTMLWindow2 GetParentWindow();
-            [return: MarshalAs(UnmanagedType.Interface)] object GetStyleSheets();
-            void SetOnbeforeupdate(object p);
-            object GetOnbeforeupdate();
-            void SetOnerrorupdate(object p);
-            object GetOnerrorupdate();
-            string toString();
-            [return: MarshalAs(UnmanagedType.Interface)] object CreateStyleSheet(string bstrHref, int lIndex);
-        }
-
-        [ComVisible(true), Guid("3050F485-98B5-11CF-BB82-00AA00BDCE0B"),
-        InterfaceType(ComInterfaceType.InterfaceIsDual)]
-        internal interface IHTMLDocument3
-        {
-            void ReleaseCapture();
-            void Recalc([In] bool fForce);
-            object CreateTextNode([In] string text);
-            [return: MarshalAs(UnmanagedType.Interface)] IHTMLElement GetDocumentElement();
-            string GetUniqueID();
-            bool AttachEvent([In] string ev, [In, MarshalAs(UnmanagedType.IDispatch)] object pdisp);
-            void DetachEvent([In] string ev, [In, MarshalAs(UnmanagedType.IDispatch)] object pdisp);
-            void SetOnrowsdelete([In] object p);
-            object GetOnrowsdelete();
-            void SetOnrowsinserted([In] object p);
-            object GetOnrowsinserted();
-            void SetOncellchange([In] object p);
-            object GetOncellchange();
-            void SetOndatasetchanged([In] object p);
-            object GetOndatasetchanged();
-            void SetOndataavailable([In] object p);
-            object GetOndataavailable();
-            void SetOndatasetcomplete([In] object p);
-            object GetOndatasetcomplete();
-            void SetOnpropertychange([In] object p);
-            object GetOnpropertychange();
-            void SetDir([In] string p);
-            string GetDir();
-            void SetOncontextmenu([In] object p);
-            object GetOncontextmenu();
-            void SetOnstop([In] object p);
-            object GetOnstop();
-            object CreateDocumentFragment();
-            object GetParentDocument();
-            void SetEnableDownload([In] bool p);
-            bool GetEnableDownload();
-            void SetBaseUrl([In] string p);
-            string GetBaseUrl();
-            [return: MarshalAs(UnmanagedType.IDispatch)] object GetChildNodes();
-            void SetInheritStyleSheets([In] bool p);
-            bool GetInheritStyleSheets();
-            void SetOnbeforeeditfocus([In] object p);
-            object GetOnbeforeeditfocus();
-            [return: MarshalAs(UnmanagedType.Interface)] IHTMLElementCollection GetElementsByName([In] string v);
-            [return: MarshalAs(UnmanagedType.Interface)] IHTMLElement GetElementById([In] string v);
-            [return: MarshalAs(UnmanagedType.Interface)] IHTMLElementCollection GetElementsByTagName([In] string v);
-        }
-
-        [ComVisible(true), Guid("3050F69A-98B5-11CF-BB82-00AA00BDCE0B"),
-        InterfaceType(System.Runtime.InteropServices.ComInterfaceType.InterfaceIsDual)]
-        internal interface IHTMLDocument4
-        {
-            void Focus();
-            bool HasFocus();
-            void SetOnselectionchange(object p);
-            object GetOnselectionchange();
-            object GetNamespaces();
-            object createDocumentFromUrl(string bstrUrl, string bstrOptions);
-            void SetMedia(string bstrMedia);
-            string GetMedia();
-            object CreateEventObject([In, Optional] ref object eventObject);
-            bool FireEvent(string eventName);
-            object CreateRenderStyle(string bstr);
-            void SetOncontrolselect(object p);
-            object GetOncontrolselect();
-            string GetURLUnencoded();
-        }
-
-        [ComImport(), Guid("3050f613-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsIDispatch),
-        TypeLibType(TypeLibTypeFlags.FHidden)]
-        public interface DHTMLDocumentEvents2
-        {
-            [DispId(-2147418102)] bool onhelp(IHTMLEventObj evtObj);
-            [DispId(-600)] bool onclick(IHTMLEventObj evtObj);
-            [DispId(-601)] bool ondblclick(IHTMLEventObj evtObj);
-            [DispId(-602)] void onkeydown(IHTMLEventObj evtObj);
-            [DispId(-604)] void onkeyup(IHTMLEventObj evtObj);
-            [DispId(-603)] bool onkeypress(IHTMLEventObj evtObj);
-            [DispId(-605)] void onmousedown(IHTMLEventObj evtObj);
-            [DispId(-606)] void onmousemove(IHTMLEventObj evtObj);
-            [DispId(-607)] void onmouseup(IHTMLEventObj evtObj);
-            [DispId(-2147418103)] void onmouseout(IHTMLEventObj evtObj);
-            [DispId(-2147418104)] void onmouseover(IHTMLEventObj evtObj);
-            [DispId(-609)] void onreadystatechange(IHTMLEventObj evtObj);
-            [DispId(-2147418108)] bool onbeforeupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418107)] void onafterupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418106)] bool onrowexit(IHTMLEventObj evtObj);
-            [DispId(-2147418105)] void onrowenter(IHTMLEventObj evtObj);
-            [DispId(-2147418101)] bool ondragstart(IHTMLEventObj evtObj);
-            [DispId(-2147418100)] bool onselectstart(IHTMLEventObj evtObj);
-            [DispId(-2147418099)] bool onerrorupdate(IHTMLEventObj evtObj);
-            [DispId(1023)] bool oncontextmenu(IHTMLEventObj evtObj);
-            [DispId(1026)] bool onstop(IHTMLEventObj evtObj);
-            [DispId(-2147418080)] void onrowsdelete(IHTMLEventObj evtObj);
-            [DispId(-2147418079)] void onrowsinserted(IHTMLEventObj evtObj);
-            [DispId(-2147418078)] void oncellchange(IHTMLEventObj evtObj);
-            [DispId(-2147418093)] void onpropertychange(IHTMLEventObj evtObj);
-            [DispId(-2147418098)] void ondatasetchanged(IHTMLEventObj evtObj);
-            [DispId(-2147418097)] void ondataavailable(IHTMLEventObj evtObj);
-            [DispId(-2147418096)] void ondatasetcomplete(IHTMLEventObj evtObj);
-            [DispId(1027)] void onbeforeeditfocus(IHTMLEventObj evtObj);
-            [DispId(1037)] void onselectionchange(IHTMLEventObj evtObj);
-            [DispId(1036)] bool oncontrolselect(IHTMLEventObj evtObj);
-            [DispId(1033)] bool onmousewheel(IHTMLEventObj evtObj);
-            [DispId(1048)] void onfocusin(IHTMLEventObj evtObj);
-            [DispId(1049)] void onfocusout(IHTMLEventObj evtObj);
-            [DispId(1044)] void onactivate(IHTMLEventObj evtObj);
-            [DispId(1045)] void ondeactivate(IHTMLEventObj evtObj);
-            [DispId(1047)] bool onbeforeactivate(IHTMLEventObj evtObj);
-            [DispId(1034)] bool onbeforedeactivate(IHTMLEventObj evtObj);
-        }
-
-        [ComVisible(true), Guid("332C4426-26CB-11D0-B483-00C04FD90119"),
-        InterfaceType(ComInterfaceType.InterfaceIsDual)]
-        internal interface IHTMLFramesCollection2
-        {
-            object Item(ref object idOrName);
-            int GetLength();
-        }
-
-        [ComVisible(true), Guid("332C4427-26CB-11D0-B483-00C04FD90119"),
-        InterfaceType(ComInterfaceType.InterfaceIsDual)]
-        public interface IHTMLWindow2
-        {
-            [return: MarshalAs(UnmanagedType.IDispatch)] object Item([In] ref object pvarIndex);
-            int GetLength();
-            [return: MarshalAs(UnmanagedType.Interface)] IHTMLFramesCollection2 GetFrames();
-            void SetDefaultStatus([In] string p);
-            string GetDefaultStatus();
-            void SetStatus([In] string p);
-            string GetStatus();
-            int SetTimeout([In] string expression, [In] int msec, [In] ref object language);
-            void ClearTimeout([In] int timerID);
-            void Alert([In] string message);
-            bool Confirm([In] string message);
-            [return: MarshalAs(UnmanagedType.Struct)] object Prompt([In] string message, [In] string defstr);
-            object GetImage();
-            [return: MarshalAs(UnmanagedType.Interface)] IHTMLLocation GetLocation();
-            [return: MarshalAs(UnmanagedType.Interface)] IOmHistory GetHistory();
-            void Close();
-            void SetOpener([In] object p);
-            [return: MarshalAs(UnmanagedType.IDispatch)] object GetOpener();
-            [return: MarshalAs(UnmanagedType.Interface)] IOmNavigator GetNavigator();
-            void SetName([In] string p);
-            string GetName();
-            [return: MarshalAs(UnmanagedType.Interface)] IHTMLWindow2 GetParent();
-            [return: MarshalAs(UnmanagedType.Interface)] IHTMLWindow2 Open([In] string URL, [In] string name, [In] string features, [In] bool replace);
-            object GetSelf();
-            object GetTop();
-            object GetWindow();
-            void Navigate([In] string URL);
-            void SetOnfocus([In] object p);
-            object GetOnfocus();
-            void SetOnblur([In] object p);
-            object GetOnblur();
-            void SetOnload([In] object p);
-            object GetOnload();
-            void SetOnbeforeunload(object p);
-            object GetOnbeforeunload();
-            void SetOnunload([In] object p);
-            object GetOnunload();
-            void SetOnhelp(object p);
-            object GetOnhelp();
-            void SetOnerror([In] object p);
-            object GetOnerror();
-            void SetOnresize([In] object p);
-            object GetOnresize();
-            void SetOnscroll([In] object p);
-            object GetOnscroll();
-            [return: MarshalAs(UnmanagedType.Interface)] IHTMLDocument2 GetDocument();
-            [return: MarshalAs(UnmanagedType.Interface)] IHTMLEventObj GetEvent();
-            object Get_newEnum();
-            object ShowModalDialog([In] string dialog, [In] ref object varArgIn, [In] ref object varOptions);
-            void ShowHelp([In] string helpURL, [In] object helpArg, [In] string features);
-            [return: MarshalAs(UnmanagedType.Interface)] IHTMLScreen GetScreen();
-            object GetOption();
-            void Focus();
-            bool GetClosed();
-            void Blur();
-            void Scroll([In] int x, [In] int y);
-            object GetClientInformation();
-            int SetInterval([In] string expression, [In] int msec, [In] ref object language);
-            void ClearInterval([In] int timerID);
-            void SetOffscreenBuffering([In] object p);
-            object GetOffscreenBuffering();
-            [return: MarshalAs(UnmanagedType.Struct)] object ExecScript([In] string code, [In] string language);
-            string toString();
-            void ScrollBy([In] int x, [In] int y);
-            void ScrollTo([In] int x, [In] int y);
-            void MoveTo([In] int x, [In] int y);
-            void MoveBy([In] int x, [In] int y);
-            void ResizeTo([In] int x, [In] int y);
-            void ResizeBy([In] int x, [In] int y);
-            object GetExternal();
-        }
-
-        [ComVisible(true), Guid("3050f4ae-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsDual)]
-        public interface IHTMLWindow3
-        {
-            int GetScreenLeft();
-            int GetScreenTop();
-            bool AttachEvent(string ev, [In, MarshalAs(UnmanagedType.IDispatch)] object pdisp);
-            void DetachEvent(string ev, [In, MarshalAs(UnmanagedType.IDispatch)] object pdisp);
-            int SetTimeout([In]ref object expression, int msec, [In] ref object language);
-            int SetInterval([In]ref object expression, int msec, [In] ref object language);
-            void Print();
-            void SetBeforePrint(object o);
-            object GetBeforePrint();
-            void SetAfterPrint(object o);
-            object GetAfterPrint();
-            object GetClipboardData();
-            object ShowModelessDialog(string url, object varArgIn, object options);
-        }
-
-        [ComVisible(true), Guid("3050f6cf-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsDual)]
-        public interface IHTMLWindow4
-        {
-            [return: MarshalAs(UnmanagedType.IDispatch)] object CreatePopup([In] ref object reserved);
-            [return: MarshalAs(UnmanagedType.Interface)] object frameElement();
-        }
-
-        [ComImport(), Guid("3050f625-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsIDispatch),
-        TypeLibType(TypeLibTypeFlags.FHidden)]
-        public interface DHTMLWindowEvents2
-        {
-            [DispId(1003)] void onload(IHTMLEventObj evtObj);
-            [DispId(1008)] void onunload(IHTMLEventObj evtObj);
-            [DispId(-2147418102)] bool onhelp(IHTMLEventObj evtObj);
-            [DispId(-2147418111)] void onfocus(IHTMLEventObj evtObj);
-            [DispId(-2147418112)] void onblur(IHTMLEventObj evtObj);
-            [DispId(1002)] bool onerror(string description, string url, int line);
-            [DispId(1016)] void onresize(IHTMLEventObj evtObj);
-            [DispId(1014)] void onscroll(IHTMLEventObj evtObj);
-            [DispId(1017)] void onbeforeunload(IHTMLEventObj evtObj);
-            [DispId(1024)] void onbeforeprint(IHTMLEventObj evtObj);
-            [DispId(1025)] void onafterprint(IHTMLEventObj evtObj);
-        }
-
-        [ComVisible(true), Guid("3050f35c-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsDual)]
-        public interface IHTMLScreen
-        {
-            int GetColorDepth();
-            void SetBufferDepth(int d);
-            int GetBufferDepth();
-            int GetWidth();
-            int GetHeight();
-            void SetUpdateInterval(int i);
-            int GetUpdateInterval();
-            int GetAvailHeight();
-            int GetAvailWidth();
-            bool GetFontSmoothingEnabled();
-        }
-
-        [ComVisible(true), Guid("163BB1E0-6E00-11CF-837A-48DC04C10000"),
-        InterfaceType(ComInterfaceType.InterfaceIsDual)]
-        internal interface IHTMLLocation
-        {
-            void SetHref([In] string p);
-            string GetHref();
-            void SetProtocol([In] string p);
-            string GetProtocol();
-            void SetHost([In] string p);
-            string GetHost();
-            void SetHostname([In] string p);
-            string GetHostname();
-            void SetPort([In] string p);
-            string GetPort();
-            void SetPathname([In] string p);
-            string GetPathname();
-            void SetSearch([In] string p);
-            string GetSearch();
-            void SetHash([In] string p);
-            string GetHash();
-            void Reload([In] bool flag);
-            void Replace([In] string bstr);
-            void Assign([In] string bstr);
-        }
-
-        [ComVisible(true), Guid("FECEAAA2-8405-11CF-8BA1-00AA00476DA6"),
-        InterfaceType(ComInterfaceType.InterfaceIsDual)]
-        internal interface IOmHistory
-        {
-            short GetLength();
-            void Back();
-            void Forward();
-            void Go([In] ref object pvargdistance);
-        }
-
-        [ComVisible(true), Guid("FECEAAA5-8405-11CF-8BA1-00AA00476DA6"),
-        InterfaceType(ComInterfaceType.InterfaceIsDual)]
-        internal interface IOmNavigator
-        {
-            string GetAppCodeName();
-            string GetAppName();
-            string GetAppVersion();
-            string GetUserAgent();
-            bool JavaEnabled();
-            bool TaintEnabled();
-            object GetMimeTypes();
-            object GetPlugins();
-            bool GetCookieEnabled();
-            object GetOpsProfile();
-            string GetCpuClass();
-            string GetSystemLanguage();
-            string GetBrowserLanguage();
-            string GetUserLanguage();
-            string GetPlatform();
-            string GetAppMinorVersion();
-            int GetConnectionSpeed();
-            bool GetOnLine();
-            object GetUserProfile();
-        }
-
-        [ComVisible(true), Guid("3050F32D-98B5-11CF-BB82-00AA00BDCE0B"),
-        InterfaceType(System.Runtime.InteropServices.ComInterfaceType.InterfaceIsDual)]
-        internal interface IHTMLEventObj
-        {
-            [return: MarshalAs(UnmanagedType.Interface)] IHTMLElement GetSrcElement();
-            bool GetAltKey();
-            bool GetCtrlKey();
-            bool GetShiftKey();
-            void SetReturnValue(object p);
-            object GetReturnValue();
-            void SetCancelBubble(bool p);
-            bool GetCancelBubble();
-            [return: MarshalAs(UnmanagedType.Interface)] IHTMLElement GetFromElement();
-            [return: MarshalAs(UnmanagedType.Interface)] IHTMLElement GetToElement();
-            void SetKeyCode([In] int p);
-            int GetKeyCode();
-            int GetButton();
-            string GetEventType();
-            string GetQualifier();
-            int GetReason();
-            int GetX();
-            int GetY();
-            int GetClientX();
-            int GetClientY();
-            int GetOffsetX();
-            int GetOffsetY();
-            int GetScreenX();
-            int GetScreenY();
-            object GetSrcFilter();
-        }
-
-        [ComVisible(true), Guid("3050F21F-98B5-11CF-BB82-00AA00BDCE0B"),
-        InterfaceType(ComInterfaceType.InterfaceIsDual)]
-        internal interface IHTMLElementCollection
-        {
-            string toString();
-            void SetLength(int p);
-            int GetLength();
-            [return: MarshalAs(UnmanagedType.Interface)] object Get_newEnum();
-            [return: MarshalAs(UnmanagedType.IDispatch)] object Item(object idOrName, object index);
-            [return: MarshalAs(UnmanagedType.Interface)] object Tags(object tagName);
-        }
-
-        [ComVisible(true), Guid("3050F1FF-98B5-11CF-BB82-00AA00BDCE0B"),
-        InterfaceType(ComInterfaceType.InterfaceIsDual)]
-        internal interface IHTMLElement
-        {
-            void SetAttribute(string attributeName, object attributeValue, int lFlags);
-            object GetAttribute(string attributeName, int lFlags);
-            bool RemoveAttribute(string strAttributeName, int lFlags);
-            void SetClassName(string p);
-            string GetClassName();
-            void SetId(string p);
-            string GetId();
-            string GetTagName();
-            IHTMLElement GetParentElement();
-            IHTMLStyle GetStyle();
-            void SetOnhelp(object p);
-            object GetOnhelp();
-            void SetOnclick(object p);
-            object GetOnclick();
-            void SetOndblclick(object p);
-            object GetOndblclick();
-            void SetOnkeydown(object p);
-            object GetOnkeydown();
-            void SetOnkeyup(object p);
-            object GetOnkeyup();
-            void SetOnkeypress(object p);
-            object GetOnkeypress();
-            void SetOnmouseout(object p);
-            object GetOnmouseout();
-            void SetOnmouseover(object p);
-            object GetOnmouseover();
-            void SetOnmousemove(object p);
-            object GetOnmousemove();
-            void SetOnmousedown(object p);
-            object GetOnmousedown();
-            void SetOnmouseup(object p);
-            object GetOnmouseup();
-            [return: MarshalAs(UnmanagedType.Interface)] IHTMLDocument2 GetDocument();
-            void SetTitle(string p);
-            string GetTitle();
-            void SetLanguage(string p);
-            string GetLanguage();
-            void SetOnselectstart(object p);
-            object GetOnselectstart();
-            void ScrollIntoView(object varargStart);
-            bool Contains(IHTMLElement pChild);
-            int GetSourceIndex();
-            object GetRecordNumber();
-            void SetLang(string p);
-            string GetLang();
-            int GetOffsetLeft();
-            int GetOffsetTop();
-            int GetOffsetWidth();
-            int GetOffsetHeight();
-            [return: MarshalAs(UnmanagedType.Interface)] IHTMLElement GetOffsetParent();
-            void SetInnerHTML(string p);
-            string GetInnerHTML();
-            void SetInnerText(string p);
-            string GetInnerText();
-            void SetOuterHTML(string p);
-            string GetOuterHTML();
-            void SetOuterText(string p);
-            string GetOuterText();
-            void InsertAdjacentHTML(string @where,
-                string html);
-            void InsertAdjacentText(string @where,
-                string text);
-            IHTMLElement GetParentTextEdit();
-            bool GetIsTextEdit();
-            void Click();
-            [return: MarshalAs(UnmanagedType.Interface)] object GetFilters();
-            void SetOndragstart(object p);
-            object GetOndragstart();
-            string toString();
-            void SetOnbeforeupdate(object p);
-            object GetOnbeforeupdate();
-            void SetOnafterupdate(object p);
-            object GetOnafterupdate();
-            void SetOnerrorupdate(object p);
-            object GetOnerrorupdate();
-            void SetOnrowexit(object p);
-            object GetOnrowexit();
-            void SetOnrowenter(object p);
-            object GetOnrowenter();
-            void SetOndatasetchanged(object p);
-            object GetOndatasetchanged();
-            void SetOndataavailable(object p);
-            object GetOndataavailable();
-            void SetOndatasetcomplete(object p);
-            object GetOndatasetcomplete();
-            void SetOnfilterchange(object p);
-            object GetOnfilterchange();
-            [return: MarshalAs(UnmanagedType.IDispatch)] object GetChildren();
-            [return: MarshalAs(UnmanagedType.IDispatch)] object GetAll();
-        }
-
-        [ComVisible(true), Guid("3050f434-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsDual)]
-        internal interface IHTMLElement2
-        {
-            string ScopeName();
-            void SetCapture(bool containerCapture);
-            void ReleaseCapture();
-            void SetOnLoseCapture(object v);
-            object GetOnLoseCapture();
-            string GetComponentFromPoint(int x, int y);
-            void DoScroll(object component);
-            void SetOnScroll(object v);
-            object GetOnScroll();
-            void SetOnDrag(object v);
-            object GetOnDrag();
-            void SetOnDragEnd(object v);
-            object GetOnDragEnd();
-            void SetOnDragEnter(object v);
-            object GetOnDragEnter();
-            void SetOnDragOver(object v);
-            object GetOnDragOver();
-            void SetOnDragleave(object v);
-            object GetOnDragLeave();
-            void SetOnDrop(object v);
-            object GetOnDrop();
-            void SetOnBeforeCut(object v);
-            object GetOnBeforeCut();
-            void SetOnCut(object v);
-            object GetOnCut();
-            void SetOnBeforeCopy(object v);
-            object GetOnBeforeCopy();
-            void SetOnCopy(object v);
-            object GetOnCopy(object p);
-            void SetOnBeforePaste(object v);
-            object GetOnBeforePaste(object p);
-            void SetOnPaste(object v);
-            object GetOnPaste(object p);
-            object GetCurrentStyle();
-            void SetOnPropertyChange(object v);
-            object GetOnPropertyChange(object p);
-            object GetClientRects();
-            object GetBoundingClientRect();
-            void SetExpression(string propName, string expression, string language);
-            object GetExpression(string propName);
-            bool RemoveExpression(string propName);
-            void SetTabIndex(int v);
-            short GetTabIndex();
-            void Focus();
-            void SetAccessKey(string v);
-            string GetAccessKey();
-            void SetOnBlur(object v);
-            object GetOnBlur();
-            void SetOnFocus(object v);
-            object GetOnFocus();
-            void SetOnResize(object v);
-            object GetOnResize();
-            void Blur();
-            void AddFilter(object pUnk);
-            void RemoveFilter(object pUnk);
-            int ClientHeight();
-            int ClientWidth();
-            int ClientTop();
-            int ClientLeft();
-            bool AttachEvent(string ev, [In, MarshalAs(UnmanagedType.IDispatch)] object pdisp);
-            void DetachEvent(string ev, [In, MarshalAs(UnmanagedType.IDispatch)] object pdisp);
-            object ReadyState();
-            void SetOnReadyStateChange(object v);
-            object GetOnReadyStateChange();
-            void SetOnRowsDelete(object v);
-            object GetOnRowsDelete();
-            void SetOnRowsInserted(object v);
-            object GetOnRowsInserted();
-            void SetOnCellChange(object v);
-            object GetOnCellChange();
-            void SetDir(string v);
-            string GetDir();
-            object CreateControlRange();
-            int GetScrollHeight();
-            int GetScrollWidth();
-            void SetScrollTop(int v);
-            int GetScrollTop();
-            void SetScrollLeft(int v);
-            int GetScrollLeft();
-            void ClearAttributes();
-            void MergeAttributes(object mergeThis);
-            void SetOnContextMenu(object v);
-            object GetOnContextMenu();
-            [return: MarshalAs(UnmanagedType.Interface)]
-            IHTMLElement
-                InsertAdjacentElement(string @where,
-                    [In, MarshalAs(UnmanagedType.Interface)] IHTMLElement insertedElement);
-            [return: MarshalAs(UnmanagedType.Interface)]
-            IHTMLElement
-                applyElement([In, MarshalAs(UnmanagedType.Interface)] IHTMLElement apply,
-                    string @where);
-            string GetAdjacentText(string @where);
-            string ReplaceAdjacentText(string @where, string newText);
-            bool CanHaveChildren();
-            int AddBehavior(string url, ref object oFactory);
-            bool RemoveBehavior(int cookie);
-            object GetRuntimeStyle();
-            object GetBehaviorUrns();
-            void SetTagUrn(string v);
-            string GetTagUrn();
-            void SetOnBeforeEditFocus(object v);
-            object GetOnBeforeEditFocus();
-            int GetReadyStateValue();
-            [return: MarshalAs(UnmanagedType.Interface)] IHTMLElementCollection GetElementsByTagName(string v);
-        }
-
-        [ComVisible(true), Guid("3050f673-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsDual)]
-        internal interface IHTMLElement3
-        {
-            void MergeAttributes(object mergeThis, object pvarFlags);
-            bool IsMultiLine();
-            bool CanHaveHTML();
-            void SetOnLayoutComplete(object v);
-            object GetOnLayoutComplete();
-            void SetOnPage(object v);
-            object GetOnPage();
-            void SetInflateBlock(bool v);
-            bool GetInflateBlock();
-            void SetOnBeforeDeactivate(object v);
-            object GetOnBeforeDeactivate();
-            void SetActive();
-            void SetContentEditable(string v);
-            string GetContentEditable();
-            bool IsContentEditable();
-            void SetHideFocus(bool v);
-            bool GetHideFocus();
-            void SetDisabled(bool v);
-            bool GetDisabled();
-            bool IsDisabled();
-            void SetOnMove(object v);
-            object GetOnMove();
-            void SetOnControlSelect(object v);
-            object GetOnControlSelect();
-            bool FireEvent(string bstrEventName, IntPtr pvarEventObject);
-            void SetOnResizeStart(object v);
-            object GetOnResizeStart();
-            void SetOnResizeEnd(object v);
-            object GetOnResizeEnd();
-            void SetOnMoveStart(object v);
-            object GetOnMoveStart();
-            void SetOnMoveEnd(object v);
-            object GetOnMoveEnd();
-            void SetOnMouseEnter(object v);
-            object GetOnMouseEnter();
-            void SetOnMouseLeave(object v);
-            object GetOnMouseLeave();
-            void SetOnActivate(object v);
-            object GetOnActivate();
-            void SetOnDeactivate(object v);
-            object GetOnDeactivate();
-            bool DragDrop();
-            int GlyphMode();
-        }
-
-        [ComVisible(true), Guid("3050f5da-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsDual)]
-        public interface IHTMLDOMNode
-        {
-            long GetNodeType();
-            IHTMLDOMNode GetParentNode();
-            bool HasChildNodes();
-            object GetChildNodes();
-            object GetAttributes();
-            IHTMLDOMNode InsertBefore(IHTMLDOMNode newChild, object refChild);
-            IHTMLDOMNode RemoveChild(IHTMLDOMNode oldChild);
-            IHTMLDOMNode ReplaceChild(IHTMLDOMNode newChild, IHTMLDOMNode oldChild);
-            IHTMLDOMNode CloneNode(bool fDeep);
-            IHTMLDOMNode RemoveNode(bool fDeep);
-            IHTMLDOMNode SwapNode(IHTMLDOMNode otherNode);
-            IHTMLDOMNode ReplaceNode(IHTMLDOMNode replacement);
-            IHTMLDOMNode AppendChild(IHTMLDOMNode newChild);
-            string NodeName();
-            void SetNodeValue(object v);
-            object GetNodeValue();
-            IHTMLDOMNode FirstChild();
-            IHTMLDOMNode LastChild();
-            IHTMLDOMNode PreviousSibling();
-            IHTMLDOMNode NextSibling();
-        };
-
-        [ComImport(), Guid("3050f60f-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsIDispatch),
-        TypeLibType(TypeLibTypeFlags.FHidden)]
-        public interface DHTMLElementEvents2
-        {
-            [DispId(-2147418102)] bool onhelp(IHTMLEventObj evtObj);
-            [DispId(-600)] bool onclick(IHTMLEventObj evtObj);
-            [DispId(-601)] bool ondblclick(IHTMLEventObj evtObj);
-            [DispId(-603)] bool onkeypress(IHTMLEventObj evtObj);
-            [DispId(-602)] void onkeydown(IHTMLEventObj evtObj);
-            [DispId(-604)] void onkeyup(IHTMLEventObj evtObj);
-            [DispId(-2147418103)] void onmouseout(IHTMLEventObj evtObj);
-            [DispId(-2147418104)] void onmouseover(IHTMLEventObj evtObj);
-            [DispId(-606)] void onmousemove(IHTMLEventObj evtObj);
-            [DispId(-605)] void onmousedown(IHTMLEventObj evtObj);
-            [DispId(-607)] void onmouseup(IHTMLEventObj evtObj);
-            [DispId(-2147418100)] bool onselectstart(IHTMLEventObj evtObj);
-            [DispId(-2147418095)] void onfilterchange(IHTMLEventObj evtObj);
-            [DispId(-2147418101)] bool ondragstart(IHTMLEventObj evtObj);
-            [DispId(-2147418108)] bool onbeforeupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418107)] void onafterupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418099)] bool onerrorupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418106)] bool onrowexit(IHTMLEventObj evtObj);
-            [DispId(-2147418105)] void onrowenter(IHTMLEventObj evtObj);
-            [DispId(-2147418098)] void ondatasetchanged(IHTMLEventObj evtObj);
-            [DispId(-2147418097)] void ondataavailable(IHTMLEventObj evtObj);
-            [DispId(-2147418096)] void ondatasetcomplete(IHTMLEventObj evtObj);
-            [DispId(-2147418094)] void onlosecapture(IHTMLEventObj evtObj);
-            [DispId(-2147418093)] void onpropertychange(IHTMLEventObj evtObj);
-            [DispId(1014)] void onscroll(IHTMLEventObj evtObj);
-            [DispId(-2147418111)] void onfocus(IHTMLEventObj evtObj);
-            [DispId(-2147418112)] void onblur(IHTMLEventObj evtObj);
-            [DispId(1016)] void onresize(IHTMLEventObj evtObj);
-            [DispId(-2147418092)] bool ondrag(IHTMLEventObj evtObj);
-            [DispId(-2147418091)] void ondragend(IHTMLEventObj evtObj);
-            [DispId(-2147418090)] bool ondragenter(IHTMLEventObj evtObj);
-            [DispId(-2147418089)] bool ondragover(IHTMLEventObj evtObj);
-            [DispId(-2147418088)] void ondragleave(IHTMLEventObj evtObj);
-            [DispId(-2147418087)] bool ondrop(IHTMLEventObj evtObj);
-            [DispId(-2147418083)] bool onbeforecut(IHTMLEventObj evtObj);
-            [DispId(-2147418086)] bool oncut(IHTMLEventObj evtObj);
-            [DispId(-2147418082)] bool onbeforecopy(IHTMLEventObj evtObj);
-            [DispId(-2147418085)] bool oncopy(IHTMLEventObj evtObj);
-            [DispId(-2147418081)] bool onbeforepaste(IHTMLEventObj evtObj);
-            [DispId(-2147418084)] bool onpaste(IHTMLEventObj evtObj);
-            [DispId(1023)] bool oncontextmenu(IHTMLEventObj evtObj);
-            [DispId(-2147418080)] void onrowsdelete(IHTMLEventObj evtObj);
-            [DispId(-2147418079)] void onrowsinserted(IHTMLEventObj evtObj);
-            [DispId(-2147418078)] void oncellchange(IHTMLEventObj evtObj);
-            [DispId(-609)] void onreadystatechange(IHTMLEventObj evtObj);
-            [DispId(1030)] void onlayoutcomplete(IHTMLEventObj evtObj);
-            [DispId(1031)] void onpage(IHTMLEventObj evtObj);
-            [DispId(1042)] void onmouseenter(IHTMLEventObj evtObj);
-            [DispId(1043)] void onmouseleave(IHTMLEventObj evtObj);
-            [DispId(1044)] void onactivate(IHTMLEventObj evtObj);
-            [DispId(1045)] void ondeactivate(IHTMLEventObj evtObj);
-            [DispId(1034)] bool onbeforedeactivate(IHTMLEventObj evtObj);
-            [DispId(1047)] bool onbeforeactivate(IHTMLEventObj evtObj);
-            [DispId(1048)] void onfocusin(IHTMLEventObj evtObj);
-            [DispId(1049)] void onfocusout(IHTMLEventObj evtObj);
-            [DispId(1035)] void onmove(IHTMLEventObj evtObj);
-            [DispId(1036)] bool oncontrolselect(IHTMLEventObj evtObj);
-            [DispId(1038)] bool onmovestart(IHTMLEventObj evtObj);
-            [DispId(1039)] void onmoveend(IHTMLEventObj evtObj);
-            [DispId(1040)] bool onresizestart(IHTMLEventObj evtObj);
-            [DispId(1041)] void onresizeend(IHTMLEventObj evtObj);
-            [DispId(1033)] bool onmousewheel(IHTMLEventObj evtObj);
-        }
-
-        [ComImport(), Guid("3050f610-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsIDispatch),
-        TypeLibType(TypeLibTypeFlags.FHidden)]
-        public interface DHTMLAnchorEvents2
-        {
-            [DispId(-2147418102)] bool onhelp(IHTMLEventObj evtObj);
-            [DispId(-600)] bool onclick(IHTMLEventObj evtObj);
-            [DispId(-601)] bool ondblclick(IHTMLEventObj evtObj);
-            [DispId(-603)] bool onkeypress(IHTMLEventObj evtObj);
-            [DispId(-602)] void onkeydown(IHTMLEventObj evtObj);
-            [DispId(-604)] void onkeyup(IHTMLEventObj evtObj);
-            [DispId(-2147418103)] void onmouseout(IHTMLEventObj evtObj);
-            [DispId(-2147418104)] void onmouseover(IHTMLEventObj evtObj);
-            [DispId(-606)] void onmousemove(IHTMLEventObj evtObj);
-            [DispId(-605)] void onmousedown(IHTMLEventObj evtObj);
-            [DispId(-607)] void onmouseup(IHTMLEventObj evtObj);
-            [DispId(-2147418100)] bool onselectstart(IHTMLEventObj evtObj);
-            [DispId(-2147418095)] void onfilterchange(IHTMLEventObj evtObj);
-            [DispId(-2147418101)] bool ondragstart(IHTMLEventObj evtObj);
-            [DispId(-2147418108)] bool onbeforeupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418107)] void onafterupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418099)] bool onerrorupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418106)] bool onrowexit(IHTMLEventObj evtObj);
-            [DispId(-2147418105)] void onrowenter(IHTMLEventObj evtObj);
-            [DispId(-2147418098)] void ondatasetchanged(IHTMLEventObj evtObj);
-            [DispId(-2147418097)] void ondataavailable(IHTMLEventObj evtObj);
-            [DispId(-2147418096)] void ondatasetcomplete(IHTMLEventObj evtObj);
-            [DispId(-2147418094)] void onlosecapture(IHTMLEventObj evtObj);
-            [DispId(-2147418093)] void onpropertychange(IHTMLEventObj evtObj);
-            [DispId(1014)] void onscroll(IHTMLEventObj evtObj);
-            [DispId(-2147418111)] void onfocus(IHTMLEventObj evtObj);
-            [DispId(-2147418112)] void onblur(IHTMLEventObj evtObj);
-            [DispId(1016)] void onresize(IHTMLEventObj evtObj);
-            [DispId(-2147418092)] bool ondrag(IHTMLEventObj evtObj);
-            [DispId(-2147418091)] void ondragend(IHTMLEventObj evtObj);
-            [DispId(-2147418090)] bool ondragenter(IHTMLEventObj evtObj);
-            [DispId(-2147418089)] bool ondragover(IHTMLEventObj evtObj);
-            [DispId(-2147418088)] void ondragleave(IHTMLEventObj evtObj);
-            [DispId(-2147418087)] bool ondrop(IHTMLEventObj evtObj);
-            [DispId(-2147418083)] bool onbeforecut(IHTMLEventObj evtObj);
-            [DispId(-2147418086)] bool oncut(IHTMLEventObj evtObj);
-            [DispId(-2147418082)] bool onbeforecopy(IHTMLEventObj evtObj);
-            [DispId(-2147418085)] bool oncopy(IHTMLEventObj evtObj);
-            [DispId(-2147418081)] bool onbeforepaste(IHTMLEventObj evtObj);
-            [DispId(-2147418084)] bool onpaste(IHTMLEventObj evtObj);
-            [DispId(1023)] bool oncontextmenu(IHTMLEventObj evtObj);
-            [DispId(-2147418080)] void onrowsdelete(IHTMLEventObj evtObj);
-            [DispId(-2147418079)] void onrowsinserted(IHTMLEventObj evtObj);
-            [DispId(-2147418078)] void oncellchange(IHTMLEventObj evtObj);
-            [DispId(-609)] void onreadystatechange(IHTMLEventObj evtObj);
-            [DispId(1030)] void onlayoutcomplete(IHTMLEventObj evtObj);
-            [DispId(1031)] void onpage(IHTMLEventObj evtObj);
-            [DispId(1042)] void onmouseenter(IHTMLEventObj evtObj);
-            [DispId(1043)] void onmouseleave(IHTMLEventObj evtObj);
-            [DispId(1044)] void onactivate(IHTMLEventObj evtObj);
-            [DispId(1045)] void ondeactivate(IHTMLEventObj evtObj);
-            [DispId(1034)] bool onbeforedeactivate(IHTMLEventObj evtObj);
-            [DispId(1047)] bool onbeforeactivate(IHTMLEventObj evtObj);
-            [DispId(1048)] void onfocusin(IHTMLEventObj evtObj);
-            [DispId(1049)] void onfocusout(IHTMLEventObj evtObj);
-            [DispId(1035)] void onmove(IHTMLEventObj evtObj);
-            [DispId(1036)] bool oncontrolselect(IHTMLEventObj evtObj);
-            [DispId(1038)] bool onmovestart(IHTMLEventObj evtObj);
-            [DispId(1039)] void onmoveend(IHTMLEventObj evtObj);
-            [DispId(1040)] bool onresizestart(IHTMLEventObj evtObj);
-            [DispId(1041)] void onresizeend(IHTMLEventObj evtObj);
-            [DispId(1033)] bool onmousewheel(IHTMLEventObj evtObj);
-        }
-
-        [ComImport(), Guid("3050f611-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsIDispatch),
-        TypeLibType(TypeLibTypeFlags.FHidden)]
-        public interface DHTMLAreaEvents2
-        {
-            [DispId(-2147418102)] bool onhelp(IHTMLEventObj evtObj);
-            [DispId(-600)] bool onclick(IHTMLEventObj evtObj);
-            [DispId(-601)] bool ondblclick(IHTMLEventObj evtObj);
-            [DispId(-603)] bool onkeypress(IHTMLEventObj evtObj);
-            [DispId(-602)] void onkeydown(IHTMLEventObj evtObj);
-            [DispId(-604)] void onkeyup(IHTMLEventObj evtObj);
-            [DispId(-2147418103)] void onmouseout(IHTMLEventObj evtObj);
-            [DispId(-2147418104)] void onmouseover(IHTMLEventObj evtObj);
-            [DispId(-606)] void onmousemove(IHTMLEventObj evtObj);
-            [DispId(-605)] void onmousedown(IHTMLEventObj evtObj);
-            [DispId(-607)] void onmouseup(IHTMLEventObj evtObj);
-            [DispId(-2147418100)] bool onselectstart(IHTMLEventObj evtObj);
-            [DispId(-2147418095)] void onfilterchange(IHTMLEventObj evtObj);
-            [DispId(-2147418101)] bool ondragstart(IHTMLEventObj evtObj);
-            [DispId(-2147418108)] bool onbeforeupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418107)] void onafterupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418099)] bool onerrorupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418106)] bool onrowexit(IHTMLEventObj evtObj);
-            [DispId(-2147418105)] void onrowenter(IHTMLEventObj evtObj);
-            [DispId(-2147418098)] void ondatasetchanged(IHTMLEventObj evtObj);
-            [DispId(-2147418097)] void ondataavailable(IHTMLEventObj evtObj);
-            [DispId(-2147418096)] void ondatasetcomplete(IHTMLEventObj evtObj);
-            [DispId(-2147418094)] void onlosecapture(IHTMLEventObj evtObj);
-            [DispId(-2147418093)] void onpropertychange(IHTMLEventObj evtObj);
-            [DispId(1014)] void onscroll(IHTMLEventObj evtObj);
-            [DispId(-2147418111)] void onfocus(IHTMLEventObj evtObj);
-            [DispId(-2147418112)] void onblur(IHTMLEventObj evtObj);
-            [DispId(1016)] void onresize(IHTMLEventObj evtObj);
-            [DispId(-2147418092)] bool ondrag(IHTMLEventObj evtObj);
-            [DispId(-2147418091)] void ondragend(IHTMLEventObj evtObj);
-            [DispId(-2147418090)] bool ondragenter(IHTMLEventObj evtObj);
-            [DispId(-2147418089)] bool ondragover(IHTMLEventObj evtObj);
-            [DispId(-2147418088)] void ondragleave(IHTMLEventObj evtObj);
-            [DispId(-2147418087)] bool ondrop(IHTMLEventObj evtObj);
-            [DispId(-2147418083)] bool onbeforecut(IHTMLEventObj evtObj);
-            [DispId(-2147418086)] bool oncut(IHTMLEventObj evtObj);
-            [DispId(-2147418082)] bool onbeforecopy(IHTMLEventObj evtObj);
-            [DispId(-2147418085)] bool oncopy(IHTMLEventObj evtObj);
-            [DispId(-2147418081)] bool onbeforepaste(IHTMLEventObj evtObj);
-            [DispId(-2147418084)] bool onpaste(IHTMLEventObj evtObj);
-            [DispId(1023)] bool oncontextmenu(IHTMLEventObj evtObj);
-            [DispId(-2147418080)] void onrowsdelete(IHTMLEventObj evtObj);
-            [DispId(-2147418079)] void onrowsinserted(IHTMLEventObj evtObj);
-            [DispId(-2147418078)] void oncellchange(IHTMLEventObj evtObj);
-            [DispId(-609)] void onreadystatechange(IHTMLEventObj evtObj);
-            [DispId(1030)] void onlayoutcomplete(IHTMLEventObj evtObj);
-            [DispId(1031)] void onpage(IHTMLEventObj evtObj);
-            [DispId(1042)] void onmouseenter(IHTMLEventObj evtObj);
-            [DispId(1043)] void onmouseleave(IHTMLEventObj evtObj);
-            [DispId(1044)] void onactivate(IHTMLEventObj evtObj);
-            [DispId(1045)] void ondeactivate(IHTMLEventObj evtObj);
-            [DispId(1034)] bool onbeforedeactivate(IHTMLEventObj evtObj);
-            [DispId(1047)] bool onbeforeactivate(IHTMLEventObj evtObj);
-            [DispId(1048)] void onfocusin(IHTMLEventObj evtObj);
-            [DispId(1049)] void onfocusout(IHTMLEventObj evtObj);
-            [DispId(1035)] void onmove(IHTMLEventObj evtObj);
-            [DispId(1036)] bool oncontrolselect(IHTMLEventObj evtObj);
-            [DispId(1038)] bool onmovestart(IHTMLEventObj evtObj);
-            [DispId(1039)] void onmoveend(IHTMLEventObj evtObj);
-            [DispId(1040)] bool onresizestart(IHTMLEventObj evtObj);
-            [DispId(1041)] void onresizeend(IHTMLEventObj evtObj);
-            [DispId(1033)] bool onmousewheel(IHTMLEventObj evtObj);
-        }
-
-        [ComImport(), Guid("3050f617-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsIDispatch),
-        TypeLibType(TypeLibTypeFlags.FHidden)]
-        public interface DHTMLButtonElementEvents2
-        {
-            [DispId(-2147418102)] bool onhelp(IHTMLEventObj evtObj);
-            [DispId(-600)] bool onclick(IHTMLEventObj evtObj);
-            [DispId(-601)] bool ondblclick(IHTMLEventObj evtObj);
-            [DispId(-603)] bool onkeypress(IHTMLEventObj evtObj);
-            [DispId(-602)] void onkeydown(IHTMLEventObj evtObj);
-            [DispId(-604)] void onkeyup(IHTMLEventObj evtObj);
-            [DispId(-2147418103)] void onmouseout(IHTMLEventObj evtObj);
-            [DispId(-2147418104)] void onmouseover(IHTMLEventObj evtObj);
-            [DispId(-606)] void onmousemove(IHTMLEventObj evtObj);
-            [DispId(-605)] void onmousedown(IHTMLEventObj evtObj);
-            [DispId(-607)] void onmouseup(IHTMLEventObj evtObj);
-            [DispId(-2147418100)] bool onselectstart(IHTMLEventObj evtObj);
-            [DispId(-2147418095)] void onfilterchange(IHTMLEventObj evtObj);
-            [DispId(-2147418101)] bool ondragstart(IHTMLEventObj evtObj);
-            [DispId(-2147418108)] bool onbeforeupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418107)] void onafterupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418099)] bool onerrorupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418106)] bool onrowexit(IHTMLEventObj evtObj);
-            [DispId(-2147418105)] void onrowenter(IHTMLEventObj evtObj);
-            [DispId(-2147418098)] void ondatasetchanged(IHTMLEventObj evtObj);
-            [DispId(-2147418097)] void ondataavailable(IHTMLEventObj evtObj);
-            [DispId(-2147418096)] void ondatasetcomplete(IHTMLEventObj evtObj);
-            [DispId(-2147418094)] void onlosecapture(IHTMLEventObj evtObj);
-            [DispId(-2147418093)] void onpropertychange(IHTMLEventObj evtObj);
-            [DispId(1014)] void onscroll(IHTMLEventObj evtObj);
-            [DispId(-2147418111)] void onfocus(IHTMLEventObj evtObj);
-            [DispId(-2147418112)] void onblur(IHTMLEventObj evtObj);
-            [DispId(1016)] void onresize(IHTMLEventObj evtObj);
-            [DispId(-2147418092)] bool ondrag(IHTMLEventObj evtObj);
-            [DispId(-2147418091)] void ondragend(IHTMLEventObj evtObj);
-            [DispId(-2147418090)] bool ondragenter(IHTMLEventObj evtObj);
-            [DispId(-2147418089)] bool ondragover(IHTMLEventObj evtObj);
-            [DispId(-2147418088)] void ondragleave(IHTMLEventObj evtObj);
-            [DispId(-2147418087)] bool ondrop(IHTMLEventObj evtObj);
-            [DispId(-2147418083)] bool onbeforecut(IHTMLEventObj evtObj);
-            [DispId(-2147418086)] bool oncut(IHTMLEventObj evtObj);
-            [DispId(-2147418082)] bool onbeforecopy(IHTMLEventObj evtObj);
-            [DispId(-2147418085)] bool oncopy(IHTMLEventObj evtObj);
-            [DispId(-2147418081)] bool onbeforepaste(IHTMLEventObj evtObj);
-            [DispId(-2147418084)] bool onpaste(IHTMLEventObj evtObj);
-            [DispId(1023)] bool oncontextmenu(IHTMLEventObj evtObj);
-            [DispId(-2147418080)] void onrowsdelete(IHTMLEventObj evtObj);
-            [DispId(-2147418079)] void onrowsinserted(IHTMLEventObj evtObj);
-            [DispId(-2147418078)] void oncellchange(IHTMLEventObj evtObj);
-            [DispId(-609)] void onreadystatechange(IHTMLEventObj evtObj);
-            [DispId(1030)] void onlayoutcomplete(IHTMLEventObj evtObj);
-            [DispId(1031)] void onpage(IHTMLEventObj evtObj);
-            [DispId(1042)] void onmouseenter(IHTMLEventObj evtObj);
-            [DispId(1043)] void onmouseleave(IHTMLEventObj evtObj);
-            [DispId(1044)] void onactivate(IHTMLEventObj evtObj);
-            [DispId(1045)] void ondeactivate(IHTMLEventObj evtObj);
-            [DispId(1034)] bool onbeforedeactivate(IHTMLEventObj evtObj);
-            [DispId(1047)] bool onbeforeactivate(IHTMLEventObj evtObj);
-            [DispId(1048)] void onfocusin(IHTMLEventObj evtObj);
-            [DispId(1049)] void onfocusout(IHTMLEventObj evtObj);
-            [DispId(1035)] void onmove(IHTMLEventObj evtObj);
-            [DispId(1036)] bool oncontrolselect(IHTMLEventObj evtObj);
-            [DispId(1038)] bool onmovestart(IHTMLEventObj evtObj);
-            [DispId(1039)] void onmoveend(IHTMLEventObj evtObj);
-            [DispId(1040)] bool onresizestart(IHTMLEventObj evtObj);
-            [DispId(1041)] void onresizeend(IHTMLEventObj evtObj);
-            [DispId(1033)] bool onmousewheel(IHTMLEventObj evtObj);
-        }
-
-        [ComImport(), Guid("3050f612-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsIDispatch),
-        TypeLibType(TypeLibTypeFlags.FHidden)]
-        public interface DHTMLControlElementEvents2
-        {
-            [DispId(-2147418102)] bool onhelp(IHTMLEventObj evtObj);
-            [DispId(-600)] bool onclick(IHTMLEventObj evtObj);
-            [DispId(-601)] bool ondblclick(IHTMLEventObj evtObj);
-            [DispId(-603)] bool onkeypress(IHTMLEventObj evtObj);
-            [DispId(-602)] void onkeydown(IHTMLEventObj evtObj);
-            [DispId(-604)] void onkeyup(IHTMLEventObj evtObj);
-            [DispId(-2147418103)] void onmouseout(IHTMLEventObj evtObj);
-            [DispId(-2147418104)] void onmouseover(IHTMLEventObj evtObj);
-            [DispId(-606)] void onmousemove(IHTMLEventObj evtObj);
-            [DispId(-605)] void onmousedown(IHTMLEventObj evtObj);
-            [DispId(-607)] void onmouseup(IHTMLEventObj evtObj);
-            [DispId(-2147418100)] bool onselectstart(IHTMLEventObj evtObj);
-            [DispId(-2147418095)] void onfilterchange(IHTMLEventObj evtObj);
-            [DispId(-2147418101)] bool ondragstart(IHTMLEventObj evtObj);
-            [DispId(-2147418108)] bool onbeforeupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418107)] void onafterupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418099)] bool onerrorupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418106)] bool onrowexit(IHTMLEventObj evtObj);
-            [DispId(-2147418105)] void onrowenter(IHTMLEventObj evtObj);
-            [DispId(-2147418098)] void ondatasetchanged(IHTMLEventObj evtObj);
-            [DispId(-2147418097)] void ondataavailable(IHTMLEventObj evtObj);
-            [DispId(-2147418096)] void ondatasetcomplete(IHTMLEventObj evtObj);
-            [DispId(-2147418094)] void onlosecapture(IHTMLEventObj evtObj);
-            [DispId(-2147418093)] void onpropertychange(IHTMLEventObj evtObj);
-            [DispId(1014)] void onscroll(IHTMLEventObj evtObj);
-            [DispId(-2147418111)] void onfocus(IHTMLEventObj evtObj);
-            [DispId(-2147418112)] void onblur(IHTMLEventObj evtObj);
-            [DispId(1016)] void onresize(IHTMLEventObj evtObj);
-            [DispId(-2147418092)] bool ondrag(IHTMLEventObj evtObj);
-            [DispId(-2147418091)] void ondragend(IHTMLEventObj evtObj);
-            [DispId(-2147418090)] bool ondragenter(IHTMLEventObj evtObj);
-            [DispId(-2147418089)] bool ondragover(IHTMLEventObj evtObj);
-            [DispId(-2147418088)] void ondragleave(IHTMLEventObj evtObj);
-            [DispId(-2147418087)] bool ondrop(IHTMLEventObj evtObj);
-            [DispId(-2147418083)] bool onbeforecut(IHTMLEventObj evtObj);
-            [DispId(-2147418086)] bool oncut(IHTMLEventObj evtObj);
-            [DispId(-2147418082)] bool onbeforecopy(IHTMLEventObj evtObj);
-            [DispId(-2147418085)] bool oncopy(IHTMLEventObj evtObj);
-            [DispId(-2147418081)] bool onbeforepaste(IHTMLEventObj evtObj);
-            [DispId(-2147418084)] bool onpaste(IHTMLEventObj evtObj);
-            [DispId(1023)] bool oncontextmenu(IHTMLEventObj evtObj);
-            [DispId(-2147418080)] void onrowsdelete(IHTMLEventObj evtObj);
-            [DispId(-2147418079)] void onrowsinserted(IHTMLEventObj evtObj);
-            [DispId(-2147418078)] void oncellchange(IHTMLEventObj evtObj);
-            [DispId(-609)] void onreadystatechange(IHTMLEventObj evtObj);
-            [DispId(1030)] void onlayoutcomplete(IHTMLEventObj evtObj);
-            [DispId(1031)] void onpage(IHTMLEventObj evtObj);
-            [DispId(1042)] void onmouseenter(IHTMLEventObj evtObj);
-            [DispId(1043)] void onmouseleave(IHTMLEventObj evtObj);
-            [DispId(1044)] void onactivate(IHTMLEventObj evtObj);
-            [DispId(1045)] void ondeactivate(IHTMLEventObj evtObj);
-            [DispId(1034)] bool onbeforedeactivate(IHTMLEventObj evtObj);
-            [DispId(1047)] bool onbeforeactivate(IHTMLEventObj evtObj);
-            [DispId(1048)] void onfocusin(IHTMLEventObj evtObj);
-            [DispId(1049)] void onfocusout(IHTMLEventObj evtObj);
-            [DispId(1035)] void onmove(IHTMLEventObj evtObj);
-            [DispId(1036)] bool oncontrolselect(IHTMLEventObj evtObj);
-            [DispId(1038)] bool onmovestart(IHTMLEventObj evtObj);
-            [DispId(1039)] void onmoveend(IHTMLEventObj evtObj);
-            [DispId(1040)] bool onresizestart(IHTMLEventObj evtObj);
-            [DispId(1041)] void onresizeend(IHTMLEventObj evtObj);
-            [DispId(1033)] bool onmousewheel(IHTMLEventObj evtObj);
-        }
-
-        [ComImport(), Guid("3050f614-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsIDispatch),
-        TypeLibType(TypeLibTypeFlags.FHidden)]
-        public interface DHTMLFormElementEvents2
-        {
-            [DispId(-2147418102)] bool onhelp(IHTMLEventObj evtObj);
-            [DispId(-600)] bool onclick(IHTMLEventObj evtObj);
-            [DispId(-601)] bool ondblclick(IHTMLEventObj evtObj);
-            [DispId(-603)] bool onkeypress(IHTMLEventObj evtObj);
-            [DispId(-602)] void onkeydown(IHTMLEventObj evtObj);
-            [DispId(-604)] void onkeyup(IHTMLEventObj evtObj);
-            [DispId(-2147418103)] void onmouseout(IHTMLEventObj evtObj);
-            [DispId(-2147418104)] void onmouseover(IHTMLEventObj evtObj);
-            [DispId(-606)] void onmousemove(IHTMLEventObj evtObj);
-            [DispId(-605)] void onmousedown(IHTMLEventObj evtObj);
-            [DispId(-607)] void onmouseup(IHTMLEventObj evtObj);
-            [DispId(-2147418100)] bool onselectstart(IHTMLEventObj evtObj);
-            [DispId(-2147418095)] void onfilterchange(IHTMLEventObj evtObj);
-            [DispId(-2147418101)] bool ondragstart(IHTMLEventObj evtObj);
-            [DispId(-2147418108)] bool onbeforeupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418107)] void onafterupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418099)] bool onerrorupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418106)] bool onrowexit(IHTMLEventObj evtObj);
-            [DispId(-2147418105)] void onrowenter(IHTMLEventObj evtObj);
-            [DispId(-2147418098)] void ondatasetchanged(IHTMLEventObj evtObj);
-            [DispId(-2147418097)] void ondataavailable(IHTMLEventObj evtObj);
-            [DispId(-2147418096)] void ondatasetcomplete(IHTMLEventObj evtObj);
-            [DispId(-2147418094)] void onlosecapture(IHTMLEventObj evtObj);
-            [DispId(-2147418093)] void onpropertychange(IHTMLEventObj evtObj);
-            [DispId(1014)] void onscroll(IHTMLEventObj evtObj);
-            [DispId(-2147418111)] void onfocus(IHTMLEventObj evtObj);
-            [DispId(-2147418112)] void onblur(IHTMLEventObj evtObj);
-            [DispId(1016)] void onresize(IHTMLEventObj evtObj);
-            [DispId(-2147418092)] bool ondrag(IHTMLEventObj evtObj);
-            [DispId(-2147418091)] void ondragend(IHTMLEventObj evtObj);
-            [DispId(-2147418090)] bool ondragenter(IHTMLEventObj evtObj);
-            [DispId(-2147418089)] bool ondragover(IHTMLEventObj evtObj);
-            [DispId(-2147418088)] void ondragleave(IHTMLEventObj evtObj);
-            [DispId(-2147418087)] bool ondrop(IHTMLEventObj evtObj);
-            [DispId(-2147418083)] bool onbeforecut(IHTMLEventObj evtObj);
-            [DispId(-2147418086)] bool oncut(IHTMLEventObj evtObj);
-            [DispId(-2147418082)] bool onbeforecopy(IHTMLEventObj evtObj);
-            [DispId(-2147418085)] bool oncopy(IHTMLEventObj evtObj);
-            [DispId(-2147418081)] bool onbeforepaste(IHTMLEventObj evtObj);
-            [DispId(-2147418084)] bool onpaste(IHTMLEventObj evtObj);
-            [DispId(1023)] bool oncontextmenu(IHTMLEventObj evtObj);
-            [DispId(-2147418080)] void onrowsdelete(IHTMLEventObj evtObj);
-            [DispId(-2147418079)] void onrowsinserted(IHTMLEventObj evtObj);
-            [DispId(-2147418078)] void oncellchange(IHTMLEventObj evtObj);
-            [DispId(-609)] void onreadystatechange(IHTMLEventObj evtObj);
-            [DispId(1030)] void onlayoutcomplete(IHTMLEventObj evtObj);
-            [DispId(1031)] void onpage(IHTMLEventObj evtObj);
-            [DispId(1042)] void onmouseenter(IHTMLEventObj evtObj);
-            [DispId(1043)] void onmouseleave(IHTMLEventObj evtObj);
-            [DispId(1044)] void onactivate(IHTMLEventObj evtObj);
-            [DispId(1045)] void ondeactivate(IHTMLEventObj evtObj);
-            [DispId(1034)] bool onbeforedeactivate(IHTMLEventObj evtObj);
-            [DispId(1047)] bool onbeforeactivate(IHTMLEventObj evtObj);
-            [DispId(1048)] void onfocusin(IHTMLEventObj evtObj);
-            [DispId(1049)] void onfocusout(IHTMLEventObj evtObj);
-            [DispId(1035)] void onmove(IHTMLEventObj evtObj);
-            [DispId(1036)] bool oncontrolselect(IHTMLEventObj evtObj);
-            [DispId(1038)] bool onmovestart(IHTMLEventObj evtObj);
-            [DispId(1039)] void onmoveend(IHTMLEventObj evtObj);
-            [DispId(1040)] bool onresizestart(IHTMLEventObj evtObj);
-            [DispId(1041)] void onresizeend(IHTMLEventObj evtObj);
-            [DispId(1033)] bool onmousewheel(IHTMLEventObj evtObj);
-            [DispId(1007)] bool onsubmit(IHTMLEventObj evtObj);
-            [DispId(1015)] bool onreset(IHTMLEventObj evtObj);
-        }
-
-        [ComImport(), Guid("3050f7ff-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsIDispatch),
-        TypeLibType(TypeLibTypeFlags.FHidden)]
-        public interface DHTMLFrameSiteEvents2
-        {
-            [DispId(-2147418102)] bool onhelp(IHTMLEventObj evtObj);
-            [DispId(-600)] bool onclick(IHTMLEventObj evtObj);
-            [DispId(-601)] bool ondblclick(IHTMLEventObj evtObj);
-            [DispId(-603)] bool onkeypress(IHTMLEventObj evtObj);
-            [DispId(-602)] void onkeydown(IHTMLEventObj evtObj);
-            [DispId(-604)] void onkeyup(IHTMLEventObj evtObj);
-            [DispId(-2147418103)] void onmouseout(IHTMLEventObj evtObj);
-            [DispId(-2147418104)] void onmouseover(IHTMLEventObj evtObj);
-            [DispId(-606)] void onmousemove(IHTMLEventObj evtObj);
-            [DispId(-605)] void onmousedown(IHTMLEventObj evtObj);
-            [DispId(-607)] void onmouseup(IHTMLEventObj evtObj);
-            [DispId(-2147418100)] bool onselectstart(IHTMLEventObj evtObj);
-            [DispId(-2147418095)] void onfilterchange(IHTMLEventObj evtObj);
-            [DispId(-2147418101)] bool ondragstart(IHTMLEventObj evtObj);
-            [DispId(-2147418108)] bool onbeforeupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418107)] void onafterupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418099)] bool onerrorupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418106)] bool onrowexit(IHTMLEventObj evtObj);
-            [DispId(-2147418105)] void onrowenter(IHTMLEventObj evtObj);
-            [DispId(-2147418098)] void ondatasetchanged(IHTMLEventObj evtObj);
-            [DispId(-2147418097)] void ondataavailable(IHTMLEventObj evtObj);
-            [DispId(-2147418096)] void ondatasetcomplete(IHTMLEventObj evtObj);
-            [DispId(-2147418094)] void onlosecapture(IHTMLEventObj evtObj);
-            [DispId(-2147418093)] void onpropertychange(IHTMLEventObj evtObj);
-            [DispId(1014)] void onscroll(IHTMLEventObj evtObj);
-            [DispId(-2147418111)] void onfocus(IHTMLEventObj evtObj);
-            [DispId(-2147418112)] void onblur(IHTMLEventObj evtObj);
-            [DispId(1016)] void onresize(IHTMLEventObj evtObj);
-            [DispId(-2147418092)] bool ondrag(IHTMLEventObj evtObj);
-            [DispId(-2147418091)] void ondragend(IHTMLEventObj evtObj);
-            [DispId(-2147418090)] bool ondragenter(IHTMLEventObj evtObj);
-            [DispId(-2147418089)] bool ondragover(IHTMLEventObj evtObj);
-            [DispId(-2147418088)] void ondragleave(IHTMLEventObj evtObj);
-            [DispId(-2147418087)] bool ondrop(IHTMLEventObj evtObj);
-            [DispId(-2147418083)] bool onbeforecut(IHTMLEventObj evtObj);
-            [DispId(-2147418086)] bool oncut(IHTMLEventObj evtObj);
-            [DispId(-2147418082)] bool onbeforecopy(IHTMLEventObj evtObj);
-            [DispId(-2147418085)] bool oncopy(IHTMLEventObj evtObj);
-            [DispId(-2147418081)] bool onbeforepaste(IHTMLEventObj evtObj);
-            [DispId(-2147418084)] bool onpaste(IHTMLEventObj evtObj);
-            [DispId(1023)] bool oncontextmenu(IHTMLEventObj evtObj);
-            [DispId(-2147418080)] void onrowsdelete(IHTMLEventObj evtObj);
-            [DispId(-2147418079)] void onrowsinserted(IHTMLEventObj evtObj);
-            [DispId(-2147418078)] void oncellchange(IHTMLEventObj evtObj);
-            [DispId(-609)] void onreadystatechange(IHTMLEventObj evtObj);
-            [DispId(1030)] void onlayoutcomplete(IHTMLEventObj evtObj);
-            [DispId(1031)] void onpage(IHTMLEventObj evtObj);
-            [DispId(1042)] void onmouseenter(IHTMLEventObj evtObj);
-            [DispId(1043)] void onmouseleave(IHTMLEventObj evtObj);
-            [DispId(1044)] void onactivate(IHTMLEventObj evtObj);
-            [DispId(1045)] void ondeactivate(IHTMLEventObj evtObj);
-            [DispId(1034)] bool onbeforedeactivate(IHTMLEventObj evtObj);
-            [DispId(1047)] bool onbeforeactivate(IHTMLEventObj evtObj);
-            [DispId(1048)] void onfocusin(IHTMLEventObj evtObj);
-            [DispId(1049)] void onfocusout(IHTMLEventObj evtObj);
-            [DispId(1035)] void onmove(IHTMLEventObj evtObj);
-            [DispId(1036)] bool oncontrolselect(IHTMLEventObj evtObj);
-            [DispId(1038)] bool onmovestart(IHTMLEventObj evtObj);
-            [DispId(1039)] void onmoveend(IHTMLEventObj evtObj);
-            [DispId(1040)] bool onresizestart(IHTMLEventObj evtObj);
-            [DispId(1041)] void onresizeend(IHTMLEventObj evtObj);
-            [DispId(1033)] bool onmousewheel(IHTMLEventObj evtObj);
-            [DispId(1003)] void onload(IHTMLEventObj evtObj);
-        }
-
-        [ComImport(), Guid("3050f616-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsIDispatch),
-        TypeLibType(TypeLibTypeFlags.FHidden)]
-        public interface DHTMLImgEvents2
-        {
-            [DispId(-2147418102)] bool onhelp(IHTMLEventObj evtObj);
-            [DispId(-600)] bool onclick(IHTMLEventObj evtObj);
-            [DispId(-601)] bool ondblclick(IHTMLEventObj evtObj);
-            [DispId(-603)] bool onkeypress(IHTMLEventObj evtObj);
-            [DispId(-602)] void onkeydown(IHTMLEventObj evtObj);
-            [DispId(-604)] void onkeyup(IHTMLEventObj evtObj);
-            [DispId(-2147418103)] void onmouseout(IHTMLEventObj evtObj);
-            [DispId(-2147418104)] void onmouseover(IHTMLEventObj evtObj);
-            [DispId(-606)] void onmousemove(IHTMLEventObj evtObj);
-            [DispId(-605)] void onmousedown(IHTMLEventObj evtObj);
-            [DispId(-607)] void onmouseup(IHTMLEventObj evtObj);
-            [DispId(-2147418100)] bool onselectstart(IHTMLEventObj evtObj);
-            [DispId(-2147418095)] void onfilterchange(IHTMLEventObj evtObj);
-            [DispId(-2147418101)] bool ondragstart(IHTMLEventObj evtObj);
-            [DispId(-2147418108)] bool onbeforeupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418107)] void onafterupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418099)] bool onerrorupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418106)] bool onrowexit(IHTMLEventObj evtObj);
-            [DispId(-2147418105)] void onrowenter(IHTMLEventObj evtObj);
-            [DispId(-2147418098)] void ondatasetchanged(IHTMLEventObj evtObj);
-            [DispId(-2147418097)] void ondataavailable(IHTMLEventObj evtObj);
-            [DispId(-2147418096)] void ondatasetcomplete(IHTMLEventObj evtObj);
-            [DispId(-2147418094)] void onlosecapture(IHTMLEventObj evtObj);
-            [DispId(-2147418093)] void onpropertychange(IHTMLEventObj evtObj);
-            [DispId(1014)] void onscroll(IHTMLEventObj evtObj);
-            [DispId(-2147418111)] void onfocus(IHTMLEventObj evtObj);
-            [DispId(-2147418112)] void onblur(IHTMLEventObj evtObj);
-            [DispId(1016)] void onresize(IHTMLEventObj evtObj);
-            [DispId(-2147418092)] bool ondrag(IHTMLEventObj evtObj);
-            [DispId(-2147418091)] void ondragend(IHTMLEventObj evtObj);
-            [DispId(-2147418090)] bool ondragenter(IHTMLEventObj evtObj);
-            [DispId(-2147418089)] bool ondragover(IHTMLEventObj evtObj);
-            [DispId(-2147418088)] void ondragleave(IHTMLEventObj evtObj);
-            [DispId(-2147418087)] bool ondrop(IHTMLEventObj evtObj);
-            [DispId(-2147418083)] bool onbeforecut(IHTMLEventObj evtObj);
-            [DispId(-2147418086)] bool oncut(IHTMLEventObj evtObj);
-            [DispId(-2147418082)] bool onbeforecopy(IHTMLEventObj evtObj);
-            [DispId(-2147418085)] bool oncopy(IHTMLEventObj evtObj);
-            [DispId(-2147418081)] bool onbeforepaste(IHTMLEventObj evtObj);
-            [DispId(-2147418084)] bool onpaste(IHTMLEventObj evtObj);
-            [DispId(1023)] bool oncontextmenu(IHTMLEventObj evtObj);
-            [DispId(-2147418080)] void onrowsdelete(IHTMLEventObj evtObj);
-            [DispId(-2147418079)] void onrowsinserted(IHTMLEventObj evtObj);
-            [DispId(-2147418078)] void oncellchange(IHTMLEventObj evtObj);
-            [DispId(-609)] void onreadystatechange(IHTMLEventObj evtObj);
-            [DispId(1030)] void onlayoutcomplete(IHTMLEventObj evtObj);
-            [DispId(1031)] void onpage(IHTMLEventObj evtObj);
-            [DispId(1042)] void onmouseenter(IHTMLEventObj evtObj);
-            [DispId(1043)] void onmouseleave(IHTMLEventObj evtObj);
-            [DispId(1044)] void onactivate(IHTMLEventObj evtObj);
-            [DispId(1045)] void ondeactivate(IHTMLEventObj evtObj);
-            [DispId(1034)] bool onbeforedeactivate(IHTMLEventObj evtObj);
-            [DispId(1047)] bool onbeforeactivate(IHTMLEventObj evtObj);
-            [DispId(1048)] void onfocusin(IHTMLEventObj evtObj);
-            [DispId(1049)] void onfocusout(IHTMLEventObj evtObj);
-            [DispId(1035)] void onmove(IHTMLEventObj evtObj);
-            [DispId(1036)] bool oncontrolselect(IHTMLEventObj evtObj);
-            [DispId(1038)] bool onmovestart(IHTMLEventObj evtObj);
-            [DispId(1039)] void onmoveend(IHTMLEventObj evtObj);
-            [DispId(1040)] bool onresizestart(IHTMLEventObj evtObj);
-            [DispId(1041)] void onresizeend(IHTMLEventObj evtObj);
-            [DispId(1033)] bool onmousewheel(IHTMLEventObj evtObj);
-            [DispId(1003)] void onload(IHTMLEventObj evtObj);
-            [DispId(1002)] void onerror(IHTMLEventObj evtObj);
-            [DispId(1000)] void onabort(IHTMLEventObj evtObj);
-        }
-
-        [ComImport(), Guid("3050f61a-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsIDispatch),
-        TypeLibType(TypeLibTypeFlags.FHidden)]
-        public interface DHTMLInputFileElementEvents2
-        {
-            [DispId(-2147418102)] bool onhelp(IHTMLEventObj evtObj);
-            [DispId(-600)] bool onclick(IHTMLEventObj evtObj);
-            [DispId(-601)] bool ondblclick(IHTMLEventObj evtObj);
-            [DispId(-603)] bool onkeypress(IHTMLEventObj evtObj);
-            [DispId(-602)] void onkeydown(IHTMLEventObj evtObj);
-            [DispId(-604)] void onkeyup(IHTMLEventObj evtObj);
-            [DispId(-2147418103)] void onmouseout(IHTMLEventObj evtObj);
-            [DispId(-2147418104)] void onmouseover(IHTMLEventObj evtObj);
-            [DispId(-606)] void onmousemove(IHTMLEventObj evtObj);
-            [DispId(-605)] void onmousedown(IHTMLEventObj evtObj);
-            [DispId(-607)] void onmouseup(IHTMLEventObj evtObj);
-            [DispId(-2147418100)] bool onselectstart(IHTMLEventObj evtObj);
-            [DispId(-2147418095)] void onfilterchange(IHTMLEventObj evtObj);
-            [DispId(-2147418101)] bool ondragstart(IHTMLEventObj evtObj);
-            [DispId(-2147418108)] bool onbeforeupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418107)] void onafterupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418099)] bool onerrorupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418106)] bool onrowexit(IHTMLEventObj evtObj);
-            [DispId(-2147418105)] void onrowenter(IHTMLEventObj evtObj);
-            [DispId(-2147418098)] void ondatasetchanged(IHTMLEventObj evtObj);
-            [DispId(-2147418097)] void ondataavailable(IHTMLEventObj evtObj);
-            [DispId(-2147418096)] void ondatasetcomplete(IHTMLEventObj evtObj);
-            [DispId(-2147418094)] void onlosecapture(IHTMLEventObj evtObj);
-            [DispId(-2147418093)] void onpropertychange(IHTMLEventObj evtObj);
-            [DispId(1014)] void onscroll(IHTMLEventObj evtObj);
-            [DispId(-2147418111)] void onfocus(IHTMLEventObj evtObj);
-            [DispId(-2147418112)] void onblur(IHTMLEventObj evtObj);
-            [DispId(1016)] void onresize(IHTMLEventObj evtObj);
-            [DispId(-2147418092)] bool ondrag(IHTMLEventObj evtObj);
-            [DispId(-2147418091)] void ondragend(IHTMLEventObj evtObj);
-            [DispId(-2147418090)] bool ondragenter(IHTMLEventObj evtObj);
-            [DispId(-2147418089)] bool ondragover(IHTMLEventObj evtObj);
-            [DispId(-2147418088)] void ondragleave(IHTMLEventObj evtObj);
-            [DispId(-2147418087)] bool ondrop(IHTMLEventObj evtObj);
-            [DispId(-2147418083)] bool onbeforecut(IHTMLEventObj evtObj);
-            [DispId(-2147418086)] bool oncut(IHTMLEventObj evtObj);
-            [DispId(-2147418082)] bool onbeforecopy(IHTMLEventObj evtObj);
-            [DispId(-2147418085)] bool oncopy(IHTMLEventObj evtObj);
-            [DispId(-2147418081)] bool onbeforepaste(IHTMLEventObj evtObj);
-            [DispId(-2147418084)] bool onpaste(IHTMLEventObj evtObj);
-            [DispId(1023)] bool oncontextmenu(IHTMLEventObj evtObj);
-            [DispId(-2147418080)] void onrowsdelete(IHTMLEventObj evtObj);
-            [DispId(-2147418079)] void onrowsinserted(IHTMLEventObj evtObj);
-            [DispId(-2147418078)] void oncellchange(IHTMLEventObj evtObj);
-            [DispId(-609)] void onreadystatechange(IHTMLEventObj evtObj);
-            [DispId(1030)] void onlayoutcomplete(IHTMLEventObj evtObj);
-            [DispId(1031)] void onpage(IHTMLEventObj evtObj);
-            [DispId(1042)] void onmouseenter(IHTMLEventObj evtObj);
-            [DispId(1043)] void onmouseleave(IHTMLEventObj evtObj);
-            [DispId(1044)] void onactivate(IHTMLEventObj evtObj);
-            [DispId(1045)] void ondeactivate(IHTMLEventObj evtObj);
-            [DispId(1034)] bool onbeforedeactivate(IHTMLEventObj evtObj);
-            [DispId(1047)] bool onbeforeactivate(IHTMLEventObj evtObj);
-            [DispId(1048)] void onfocusin(IHTMLEventObj evtObj);
-            [DispId(1049)] void onfocusout(IHTMLEventObj evtObj);
-            [DispId(1035)] void onmove(IHTMLEventObj evtObj);
-            [DispId(1036)] bool oncontrolselect(IHTMLEventObj evtObj);
-            [DispId(1038)] bool onmovestart(IHTMLEventObj evtObj);
-            [DispId(1039)] void onmoveend(IHTMLEventObj evtObj);
-            [DispId(1040)] bool onresizestart(IHTMLEventObj evtObj);
-            [DispId(1041)] void onresizeend(IHTMLEventObj evtObj);
-            [DispId(1033)] bool onmousewheel(IHTMLEventObj evtObj);
-            [DispId(-2147412082)] bool onchange(IHTMLEventObj evtObj);
-            [DispId(-2147412102)] void onselect(IHTMLEventObj evtObj);
-            [DispId(1003)] void onload(IHTMLEventObj evtObj);
-            [DispId(1002)] void onerror(IHTMLEventObj evtObj);
-            [DispId(1000)] void onabort(IHTMLEventObj evtObj);
-        }
-
-        [ComImport(), Guid("3050f61b-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsIDispatch),
-        TypeLibType(TypeLibTypeFlags.FHidden)]
-        public interface DHTMLInputImageEvents2
-        {
-            [DispId(-2147418102)] bool onhelp(IHTMLEventObj evtObj);
-            [DispId(-600)] bool onclick(IHTMLEventObj evtObj);
-            [DispId(-601)] bool ondblclick(IHTMLEventObj evtObj);
-            [DispId(-603)] bool onkeypress(IHTMLEventObj evtObj);
-            [DispId(-602)] void onkeydown(IHTMLEventObj evtObj);
-            [DispId(-604)] void onkeyup(IHTMLEventObj evtObj);
-            [DispId(-2147418103)] void onmouseout(IHTMLEventObj evtObj);
-            [DispId(-2147418104)] void onmouseover(IHTMLEventObj evtObj);
-            [DispId(-606)] void onmousemove(IHTMLEventObj evtObj);
-            [DispId(-605)] void onmousedown(IHTMLEventObj evtObj);
-            [DispId(-607)] void onmouseup(IHTMLEventObj evtObj);
-            [DispId(-2147418100)] bool onselectstart(IHTMLEventObj evtObj);
-            [DispId(-2147418095)] void onfilterchange(IHTMLEventObj evtObj);
-            [DispId(-2147418101)] bool ondragstart(IHTMLEventObj evtObj);
-            [DispId(-2147418108)] bool onbeforeupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418107)] void onafterupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418099)] bool onerrorupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418106)] bool onrowexit(IHTMLEventObj evtObj);
-            [DispId(-2147418105)] void onrowenter(IHTMLEventObj evtObj);
-            [DispId(-2147418098)] void ondatasetchanged(IHTMLEventObj evtObj);
-            [DispId(-2147418097)] void ondataavailable(IHTMLEventObj evtObj);
-            [DispId(-2147418096)] void ondatasetcomplete(IHTMLEventObj evtObj);
-            [DispId(-2147418094)] void onlosecapture(IHTMLEventObj evtObj);
-            [DispId(-2147418093)] void onpropertychange(IHTMLEventObj evtObj);
-            [DispId(1014)] void onscroll(IHTMLEventObj evtObj);
-            [DispId(-2147418111)] void onfocus(IHTMLEventObj evtObj);
-            [DispId(-2147418112)] void onblur(IHTMLEventObj evtObj);
-            [DispId(1016)] void onresize(IHTMLEventObj evtObj);
-            [DispId(-2147418092)] bool ondrag(IHTMLEventObj evtObj);
-            [DispId(-2147418091)] void ondragend(IHTMLEventObj evtObj);
-            [DispId(-2147418090)] bool ondragenter(IHTMLEventObj evtObj);
-            [DispId(-2147418089)] bool ondragover(IHTMLEventObj evtObj);
-            [DispId(-2147418088)] void ondragleave(IHTMLEventObj evtObj);
-            [DispId(-2147418087)] bool ondrop(IHTMLEventObj evtObj);
-            [DispId(-2147418083)] bool onbeforecut(IHTMLEventObj evtObj);
-            [DispId(-2147418086)] bool oncut(IHTMLEventObj evtObj);
-            [DispId(-2147418082)] bool onbeforecopy(IHTMLEventObj evtObj);
-            [DispId(-2147418085)] bool oncopy(IHTMLEventObj evtObj);
-            [DispId(-2147418081)] bool onbeforepaste(IHTMLEventObj evtObj);
-            [DispId(-2147418084)] bool onpaste(IHTMLEventObj evtObj);
-            [DispId(1023)] bool oncontextmenu(IHTMLEventObj evtObj);
-            [DispId(-2147418080)] void onrowsdelete(IHTMLEventObj evtObj);
-            [DispId(-2147418079)] void onrowsinserted(IHTMLEventObj evtObj);
-            [DispId(-2147418078)] void oncellchange(IHTMLEventObj evtObj);
-            [DispId(-609)] void onreadystatechange(IHTMLEventObj evtObj);
-            [DispId(1030)] void onlayoutcomplete(IHTMLEventObj evtObj);
-            [DispId(1031)] void onpage(IHTMLEventObj evtObj);
-            [DispId(1042)] void onmouseenter(IHTMLEventObj evtObj);
-            [DispId(1043)] void onmouseleave(IHTMLEventObj evtObj);
-            [DispId(1044)] void onactivate(IHTMLEventObj evtObj);
-            [DispId(1045)] void ondeactivate(IHTMLEventObj evtObj);
-            [DispId(1034)] bool onbeforedeactivate(IHTMLEventObj evtObj);
-            [DispId(1047)] bool onbeforeactivate(IHTMLEventObj evtObj);
-            [DispId(1048)] void onfocusin(IHTMLEventObj evtObj);
-            [DispId(1049)] void onfocusout(IHTMLEventObj evtObj);
-            [DispId(1035)] void onmove(IHTMLEventObj evtObj);
-            [DispId(1036)] bool oncontrolselect(IHTMLEventObj evtObj);
-            [DispId(1038)] bool onmovestart(IHTMLEventObj evtObj);
-            [DispId(1039)] void onmoveend(IHTMLEventObj evtObj);
-            [DispId(1040)] bool onresizestart(IHTMLEventObj evtObj);
-            [DispId(1041)] void onresizeend(IHTMLEventObj evtObj);
-            [DispId(1033)] bool onmousewheel(IHTMLEventObj evtObj);
-            [DispId(-2147412080)] void onload(IHTMLEventObj evtObj);
-            [DispId(-2147412083)] void onerror(IHTMLEventObj evtObj);
-            [DispId(-2147412084)] void onabort(IHTMLEventObj evtObj);
-        }
-
-        [ComImport(), Guid("3050f618-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsIDispatch),
-        TypeLibType(TypeLibTypeFlags.FHidden)]
-        public interface DHTMLInputTextElementEvents2
-        {
-            [DispId(-2147418102)] bool onhelp(IHTMLEventObj evtObj);
-            [DispId(-600)] bool onclick(IHTMLEventObj evtObj);
-            [DispId(-601)] bool ondblclick(IHTMLEventObj evtObj);
-            [DispId(-603)] bool onkeypress(IHTMLEventObj evtObj);
-            [DispId(-602)] void onkeydown(IHTMLEventObj evtObj);
-            [DispId(-604)] void onkeyup(IHTMLEventObj evtObj);
-            [DispId(-2147418103)] void onmouseout(IHTMLEventObj evtObj);
-            [DispId(-2147418104)] void onmouseover(IHTMLEventObj evtObj);
-            [DispId(-606)] void onmousemove(IHTMLEventObj evtObj);
-            [DispId(-605)] void onmousedown(IHTMLEventObj evtObj);
-            [DispId(-607)] void onmouseup(IHTMLEventObj evtObj);
-            [DispId(-2147418100)] bool onselectstart(IHTMLEventObj evtObj);
-            [DispId(-2147418095)] void onfilterchange(IHTMLEventObj evtObj);
-            [DispId(-2147418101)] bool ondragstart(IHTMLEventObj evtObj);
-            [DispId(-2147418108)] bool onbeforeupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418107)] void onafterupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418099)] bool onerrorupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418106)] bool onrowexit(IHTMLEventObj evtObj);
-            [DispId(-2147418105)] void onrowenter(IHTMLEventObj evtObj);
-            [DispId(-2147418098)] void ondatasetchanged(IHTMLEventObj evtObj);
-            [DispId(-2147418097)] void ondataavailable(IHTMLEventObj evtObj);
-            [DispId(-2147418096)] void ondatasetcomplete(IHTMLEventObj evtObj);
-            [DispId(-2147418094)] void onlosecapture(IHTMLEventObj evtObj);
-            [DispId(-2147418093)] void onpropertychange(IHTMLEventObj evtObj);
-            [DispId(1014)] void onscroll(IHTMLEventObj evtObj);
-            [DispId(-2147418111)] void onfocus(IHTMLEventObj evtObj);
-            [DispId(-2147418112)] void onblur(IHTMLEventObj evtObj);
-            [DispId(1016)] void onresize(IHTMLEventObj evtObj);
-            [DispId(-2147418092)] bool ondrag(IHTMLEventObj evtObj);
-            [DispId(-2147418091)] void ondragend(IHTMLEventObj evtObj);
-            [DispId(-2147418090)] bool ondragenter(IHTMLEventObj evtObj);
-            [DispId(-2147418089)] bool ondragover(IHTMLEventObj evtObj);
-            [DispId(-2147418088)] void ondragleave(IHTMLEventObj evtObj);
-            [DispId(-2147418087)] bool ondrop(IHTMLEventObj evtObj);
-            [DispId(-2147418083)] bool onbeforecut(IHTMLEventObj evtObj);
-            [DispId(-2147418086)] bool oncut(IHTMLEventObj evtObj);
-            [DispId(-2147418082)] bool onbeforecopy(IHTMLEventObj evtObj);
-            [DispId(-2147418085)] bool oncopy(IHTMLEventObj evtObj);
-            [DispId(-2147418081)] bool onbeforepaste(IHTMLEventObj evtObj);
-            [DispId(-2147418084)] bool onpaste(IHTMLEventObj evtObj);
-            [DispId(1023)] bool oncontextmenu(IHTMLEventObj evtObj);
-            [DispId(-2147418080)] void onrowsdelete(IHTMLEventObj evtObj);
-            [DispId(-2147418079)] void onrowsinserted(IHTMLEventObj evtObj);
-            [DispId(-2147418078)] void oncellchange(IHTMLEventObj evtObj);
-            [DispId(-609)] void onreadystatechange(IHTMLEventObj evtObj);
-            [DispId(1030)] void onlayoutcomplete(IHTMLEventObj evtObj);
-            [DispId(1031)] void onpage(IHTMLEventObj evtObj);
-            [DispId(1042)] void onmouseenter(IHTMLEventObj evtObj);
-            [DispId(1043)] void onmouseleave(IHTMLEventObj evtObj);
-            [DispId(1044)] void onactivate(IHTMLEventObj evtObj);
-            [DispId(1045)] void ondeactivate(IHTMLEventObj evtObj);
-            [DispId(1034)] bool onbeforedeactivate(IHTMLEventObj evtObj);
-            [DispId(1047)] bool onbeforeactivate(IHTMLEventObj evtObj);
-            [DispId(1048)] void onfocusin(IHTMLEventObj evtObj);
-            [DispId(1049)] void onfocusout(IHTMLEventObj evtObj);
-            [DispId(1035)] void onmove(IHTMLEventObj evtObj);
-            [DispId(1036)] bool oncontrolselect(IHTMLEventObj evtObj);
-            [DispId(1038)] bool onmovestart(IHTMLEventObj evtObj);
-            [DispId(1039)] void onmoveend(IHTMLEventObj evtObj);
-            [DispId(1040)] bool onresizestart(IHTMLEventObj evtObj);
-            [DispId(1041)] void onresizeend(IHTMLEventObj evtObj);
-            [DispId(1033)] bool onmousewheel(IHTMLEventObj evtObj);
-            [DispId(1001)] bool onchange(IHTMLEventObj evtObj);
-            [DispId(1006)] void onselect(IHTMLEventObj evtObj);
-            [DispId(1003)] void onload(IHTMLEventObj evtObj);
-            [DispId(1002)] void onerror(IHTMLEventObj evtObj);
-            [DispId(1001)] void onabort(IHTMLEventObj evtObj);
-        }
-
-        [ComImport(), Guid("3050f61c-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsIDispatch),
-        TypeLibType(TypeLibTypeFlags.FHidden)]
-        public interface DHTMLLabelEvents2
-        {
-            [DispId(-2147418102)] bool onhelp(IHTMLEventObj evtObj);
-            [DispId(-600)] bool onclick(IHTMLEventObj evtObj);
-            [DispId(-601)] bool ondblclick(IHTMLEventObj evtObj);
-            [DispId(-603)] bool onkeypress(IHTMLEventObj evtObj);
-            [DispId(-602)] void onkeydown(IHTMLEventObj evtObj);
-            [DispId(-604)] void onkeyup(IHTMLEventObj evtObj);
-            [DispId(-2147418103)] void onmouseout(IHTMLEventObj evtObj);
-            [DispId(-2147418104)] void onmouseover(IHTMLEventObj evtObj);
-            [DispId(-606)] void onmousemove(IHTMLEventObj evtObj);
-            [DispId(-605)] void onmousedown(IHTMLEventObj evtObj);
-            [DispId(-607)] void onmouseup(IHTMLEventObj evtObj);
-            [DispId(-2147418100)] bool onselectstart(IHTMLEventObj evtObj);
-            [DispId(-2147418095)] void onfilterchange(IHTMLEventObj evtObj);
-            [DispId(-2147418101)] bool ondragstart(IHTMLEventObj evtObj);
-            [DispId(-2147418108)] bool onbeforeupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418107)] void onafterupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418099)] bool onerrorupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418106)] bool onrowexit(IHTMLEventObj evtObj);
-            [DispId(-2147418105)] void onrowenter(IHTMLEventObj evtObj);
-            [DispId(-2147418098)] void ondatasetchanged(IHTMLEventObj evtObj);
-            [DispId(-2147418097)] void ondataavailable(IHTMLEventObj evtObj);
-            [DispId(-2147418096)] void ondatasetcomplete(IHTMLEventObj evtObj);
-            [DispId(-2147418094)] void onlosecapture(IHTMLEventObj evtObj);
-            [DispId(-2147418093)] void onpropertychange(IHTMLEventObj evtObj);
-            [DispId(1014)] void onscroll(IHTMLEventObj evtObj);
-            [DispId(-2147418111)] void onfocus(IHTMLEventObj evtObj);
-            [DispId(-2147418112)] void onblur(IHTMLEventObj evtObj);
-            [DispId(1016)] void onresize(IHTMLEventObj evtObj);
-            [DispId(-2147418092)] bool ondrag(IHTMLEventObj evtObj);
-            [DispId(-2147418091)] void ondragend(IHTMLEventObj evtObj);
-            [DispId(-2147418090)] bool ondragenter(IHTMLEventObj evtObj);
-            [DispId(-2147418089)] bool ondragover(IHTMLEventObj evtObj);
-            [DispId(-2147418088)] void ondragleave(IHTMLEventObj evtObj);
-            [DispId(-2147418087)] bool ondrop(IHTMLEventObj evtObj);
-            [DispId(-2147418083)] bool onbeforecut(IHTMLEventObj evtObj);
-            [DispId(-2147418086)] bool oncut(IHTMLEventObj evtObj);
-            [DispId(-2147418082)] bool onbeforecopy(IHTMLEventObj evtObj);
-            [DispId(-2147418085)] bool oncopy(IHTMLEventObj evtObj);
-            [DispId(-2147418081)] bool onbeforepaste(IHTMLEventObj evtObj);
-            [DispId(-2147418084)] bool onpaste(IHTMLEventObj evtObj);
-            [DispId(1023)] bool oncontextmenu(IHTMLEventObj evtObj);
-            [DispId(-2147418080)] void onrowsdelete(IHTMLEventObj evtObj);
-            [DispId(-2147418079)] void onrowsinserted(IHTMLEventObj evtObj);
-            [DispId(-2147418078)] void oncellchange(IHTMLEventObj evtObj);
-            [DispId(-609)] void onreadystatechange(IHTMLEventObj evtObj);
-            [DispId(1030)] void onlayoutcomplete(IHTMLEventObj evtObj);
-            [DispId(1031)] void onpage(IHTMLEventObj evtObj);
-            [DispId(1042)] void onmouseenter(IHTMLEventObj evtObj);
-            [DispId(1043)] void onmouseleave(IHTMLEventObj evtObj);
-            [DispId(1044)] void onactivate(IHTMLEventObj evtObj);
-            [DispId(1045)] void ondeactivate(IHTMLEventObj evtObj);
-            [DispId(1034)] bool onbeforedeactivate(IHTMLEventObj evtObj);
-            [DispId(1047)] bool onbeforeactivate(IHTMLEventObj evtObj);
-            [DispId(1048)] void onfocusin(IHTMLEventObj evtObj);
-            [DispId(1049)] void onfocusout(IHTMLEventObj evtObj);
-            [DispId(1035)] void onmove(IHTMLEventObj evtObj);
-            [DispId(1036)] bool oncontrolselect(IHTMLEventObj evtObj);
-            [DispId(1038)] bool onmovestart(IHTMLEventObj evtObj);
-            [DispId(1039)] void onmoveend(IHTMLEventObj evtObj);
-            [DispId(1040)] bool onresizestart(IHTMLEventObj evtObj);
-            [DispId(1041)] void onresizeend(IHTMLEventObj evtObj);
-            [DispId(1033)] bool onmousewheel(IHTMLEventObj evtObj);
-        }
-
-        [ComImport(), Guid("3050f61d-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsIDispatch),
-        TypeLibType(TypeLibTypeFlags.FHidden)]
-        public interface DHTMLLinkElementEvents2
-        {
-            [DispId(-2147418102)] bool onhelp(IHTMLEventObj evtObj);
-            [DispId(-600)] bool onclick(IHTMLEventObj evtObj);
-            [DispId(-601)] bool ondblclick(IHTMLEventObj evtObj);
-            [DispId(-603)] bool onkeypress(IHTMLEventObj evtObj);
-            [DispId(-602)] void onkeydown(IHTMLEventObj evtObj);
-            [DispId(-604)] void onkeyup(IHTMLEventObj evtObj);
-            [DispId(-2147418103)] void onmouseout(IHTMLEventObj evtObj);
-            [DispId(-2147418104)] void onmouseover(IHTMLEventObj evtObj);
-            [DispId(-606)] void onmousemove(IHTMLEventObj evtObj);
-            [DispId(-605)] void onmousedown(IHTMLEventObj evtObj);
-            [DispId(-607)] void onmouseup(IHTMLEventObj evtObj);
-            [DispId(-2147418100)] bool onselectstart(IHTMLEventObj evtObj);
-            [DispId(-2147418095)] void onfilterchange(IHTMLEventObj evtObj);
-            [DispId(-2147418101)] bool ondragstart(IHTMLEventObj evtObj);
-            [DispId(-2147418108)] bool onbeforeupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418107)] void onafterupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418099)] bool onerrorupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418106)] bool onrowexit(IHTMLEventObj evtObj);
-            [DispId(-2147418105)] void onrowenter(IHTMLEventObj evtObj);
-            [DispId(-2147418098)] void ondatasetchanged(IHTMLEventObj evtObj);
-            [DispId(-2147418097)] void ondataavailable(IHTMLEventObj evtObj);
-            [DispId(-2147418096)] void ondatasetcomplete(IHTMLEventObj evtObj);
-            [DispId(-2147418094)] void onlosecapture(IHTMLEventObj evtObj);
-            [DispId(-2147418093)] void onpropertychange(IHTMLEventObj evtObj);
-            [DispId(1014)] void onscroll(IHTMLEventObj evtObj);
-            [DispId(-2147418111)] void onfocus(IHTMLEventObj evtObj);
-            [DispId(-2147418112)] void onblur(IHTMLEventObj evtObj);
-            [DispId(1016)] void onresize(IHTMLEventObj evtObj);
-            [DispId(-2147418092)] bool ondrag(IHTMLEventObj evtObj);
-            [DispId(-2147418091)] void ondragend(IHTMLEventObj evtObj);
-            [DispId(-2147418090)] bool ondragenter(IHTMLEventObj evtObj);
-            [DispId(-2147418089)] bool ondragover(IHTMLEventObj evtObj);
-            [DispId(-2147418088)] void ondragleave(IHTMLEventObj evtObj);
-            [DispId(-2147418087)] bool ondrop(IHTMLEventObj evtObj);
-            [DispId(-2147418083)] bool onbeforecut(IHTMLEventObj evtObj);
-            [DispId(-2147418086)] bool oncut(IHTMLEventObj evtObj);
-            [DispId(-2147418082)] bool onbeforecopy(IHTMLEventObj evtObj);
-            [DispId(-2147418085)] bool oncopy(IHTMLEventObj evtObj);
-            [DispId(-2147418081)] bool onbeforepaste(IHTMLEventObj evtObj);
-            [DispId(-2147418084)] bool onpaste(IHTMLEventObj evtObj);
-            [DispId(1023)] bool oncontextmenu(IHTMLEventObj evtObj);
-            [DispId(-2147418080)] void onrowsdelete(IHTMLEventObj evtObj);
-            [DispId(-2147418079)] void onrowsinserted(IHTMLEventObj evtObj);
-            [DispId(-2147418078)] void oncellchange(IHTMLEventObj evtObj);
-            [DispId(-609)] void onreadystatechange(IHTMLEventObj evtObj);
-            [DispId(1030)] void onlayoutcomplete(IHTMLEventObj evtObj);
-            [DispId(1031)] void onpage(IHTMLEventObj evtObj);
-            [DispId(1042)] void onmouseenter(IHTMLEventObj evtObj);
-            [DispId(1043)] void onmouseleave(IHTMLEventObj evtObj);
-            [DispId(1044)] void onactivate(IHTMLEventObj evtObj);
-            [DispId(1045)] void ondeactivate(IHTMLEventObj evtObj);
-            [DispId(1034)] bool onbeforedeactivate(IHTMLEventObj evtObj);
-            [DispId(1047)] bool onbeforeactivate(IHTMLEventObj evtObj);
-            [DispId(1048)] void onfocusin(IHTMLEventObj evtObj);
-            [DispId(1049)] void onfocusout(IHTMLEventObj evtObj);
-            [DispId(1035)] void onmove(IHTMLEventObj evtObj);
-            [DispId(1036)] bool oncontrolselect(IHTMLEventObj evtObj);
-            [DispId(1038)] bool onmovestart(IHTMLEventObj evtObj);
-            [DispId(1039)] void onmoveend(IHTMLEventObj evtObj);
-            [DispId(1040)] bool onresizestart(IHTMLEventObj evtObj);
-            [DispId(1041)] void onresizeend(IHTMLEventObj evtObj);
-            [DispId(1033)] bool onmousewheel(IHTMLEventObj evtObj);
-            [DispId(-2147412080)] void onload(IHTMLEventObj evtObj);
-            [DispId(-2147412083)] void onerror(IHTMLEventObj evtObj);
-        }
-
-        [ComImport(), Guid("3050f61e-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsIDispatch),
-        TypeLibType(TypeLibTypeFlags.FHidden)]
-        public interface DHTMLMapEvents2
-        {
-            [DispId(-2147418102)] bool onhelp(IHTMLEventObj evtObj);
-            [DispId(-600)] bool onclick(IHTMLEventObj evtObj);
-            [DispId(-601)] bool ondblclick(IHTMLEventObj evtObj);
-            [DispId(-603)] bool onkeypress(IHTMLEventObj evtObj);
-            [DispId(-602)] void onkeydown(IHTMLEventObj evtObj);
-            [DispId(-604)] void onkeyup(IHTMLEventObj evtObj);
-            [DispId(-2147418103)] void onmouseout(IHTMLEventObj evtObj);
-            [DispId(-2147418104)] void onmouseover(IHTMLEventObj evtObj);
-            [DispId(-606)] void onmousemove(IHTMLEventObj evtObj);
-            [DispId(-605)] void onmousedown(IHTMLEventObj evtObj);
-            [DispId(-607)] void onmouseup(IHTMLEventObj evtObj);
-            [DispId(-2147418100)] bool onselectstart(IHTMLEventObj evtObj);
-            [DispId(-2147418095)] void onfilterchange(IHTMLEventObj evtObj);
-            [DispId(-2147418101)] bool ondragstart(IHTMLEventObj evtObj);
-            [DispId(-2147418108)] bool onbeforeupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418107)] void onafterupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418099)] bool onerrorupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418106)] bool onrowexit(IHTMLEventObj evtObj);
-            [DispId(-2147418105)] void onrowenter(IHTMLEventObj evtObj);
-            [DispId(-2147418098)] void ondatasetchanged(IHTMLEventObj evtObj);
-            [DispId(-2147418097)] void ondataavailable(IHTMLEventObj evtObj);
-            [DispId(-2147418096)] void ondatasetcomplete(IHTMLEventObj evtObj);
-            [DispId(-2147418094)] void onlosecapture(IHTMLEventObj evtObj);
-            [DispId(-2147418093)] void onpropertychange(IHTMLEventObj evtObj);
-            [DispId(1014)] void onscroll(IHTMLEventObj evtObj);
-            [DispId(-2147418111)] void onfocus(IHTMLEventObj evtObj);
-            [DispId(-2147418112)] void onblur(IHTMLEventObj evtObj);
-            [DispId(1016)] void onresize(IHTMLEventObj evtObj);
-            [DispId(-2147418092)] bool ondrag(IHTMLEventObj evtObj);
-            [DispId(-2147418091)] void ondragend(IHTMLEventObj evtObj);
-            [DispId(-2147418090)] bool ondragenter(IHTMLEventObj evtObj);
-            [DispId(-2147418089)] bool ondragover(IHTMLEventObj evtObj);
-            [DispId(-2147418088)] void ondragleave(IHTMLEventObj evtObj);
-            [DispId(-2147418087)] bool ondrop(IHTMLEventObj evtObj);
-            [DispId(-2147418083)] bool onbeforecut(IHTMLEventObj evtObj);
-            [DispId(-2147418086)] bool oncut(IHTMLEventObj evtObj);
-            [DispId(-2147418082)] bool onbeforecopy(IHTMLEventObj evtObj);
-            [DispId(-2147418085)] bool oncopy(IHTMLEventObj evtObj);
-            [DispId(-2147418081)] bool onbeforepaste(IHTMLEventObj evtObj);
-            [DispId(-2147418084)] bool onpaste(IHTMLEventObj evtObj);
-            [DispId(1023)] bool oncontextmenu(IHTMLEventObj evtObj);
-            [DispId(-2147418080)] void onrowsdelete(IHTMLEventObj evtObj);
-            [DispId(-2147418079)] void onrowsinserted(IHTMLEventObj evtObj);
-            [DispId(-2147418078)] void oncellchange(IHTMLEventObj evtObj);
-            [DispId(-609)] void onreadystatechange(IHTMLEventObj evtObj);
-            [DispId(1030)] void onlayoutcomplete(IHTMLEventObj evtObj);
-            [DispId(1031)] void onpage(IHTMLEventObj evtObj);
-            [DispId(1042)] void onmouseenter(IHTMLEventObj evtObj);
-            [DispId(1043)] void onmouseleave(IHTMLEventObj evtObj);
-            [DispId(1044)] void onactivate(IHTMLEventObj evtObj);
-            [DispId(1045)] void ondeactivate(IHTMLEventObj evtObj);
-            [DispId(1034)] bool onbeforedeactivate(IHTMLEventObj evtObj);
-            [DispId(1047)] bool onbeforeactivate(IHTMLEventObj evtObj);
-            [DispId(1048)] void onfocusin(IHTMLEventObj evtObj);
-            [DispId(1049)] void onfocusout(IHTMLEventObj evtObj);
-            [DispId(1035)] void onmove(IHTMLEventObj evtObj);
-            [DispId(1036)] bool oncontrolselect(IHTMLEventObj evtObj);
-            [DispId(1038)] bool onmovestart(IHTMLEventObj evtObj);
-            [DispId(1039)] void onmoveend(IHTMLEventObj evtObj);
-            [DispId(1040)] bool onresizestart(IHTMLEventObj evtObj);
-            [DispId(1041)] void onresizeend(IHTMLEventObj evtObj);
-            [DispId(1033)] bool onmousewheel(IHTMLEventObj evtObj);
-        }
-
-        [ComImport(), Guid("3050f61f-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsIDispatch),
-        TypeLibType(TypeLibTypeFlags.FHidden)]
-        public interface DHTMLMarqueeElementEvents2
-        {
-            [DispId(-2147418102)] bool onhelp(IHTMLEventObj evtObj);
-            [DispId(-600)] bool onclick(IHTMLEventObj evtObj);
-            [DispId(-601)] bool ondblclick(IHTMLEventObj evtObj);
-            [DispId(-603)] bool onkeypress(IHTMLEventObj evtObj);
-            [DispId(-602)] void onkeydown(IHTMLEventObj evtObj);
-            [DispId(-604)] void onkeyup(IHTMLEventObj evtObj);
-            [DispId(-2147418103)] void onmouseout(IHTMLEventObj evtObj);
-            [DispId(-2147418104)] void onmouseover(IHTMLEventObj evtObj);
-            [DispId(-606)] void onmousemove(IHTMLEventObj evtObj);
-            [DispId(-605)] void onmousedown(IHTMLEventObj evtObj);
-            [DispId(-607)] void onmouseup(IHTMLEventObj evtObj);
-            [DispId(-2147418100)] bool onselectstart(IHTMLEventObj evtObj);
-            [DispId(-2147418095)] void onfilterchange(IHTMLEventObj evtObj);
-            [DispId(-2147418101)] bool ondragstart(IHTMLEventObj evtObj);
-            [DispId(-2147418108)] bool onbeforeupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418107)] void onafterupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418099)] bool onerrorupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418106)] bool onrowexit(IHTMLEventObj evtObj);
-            [DispId(-2147418105)] void onrowenter(IHTMLEventObj evtObj);
-            [DispId(-2147418098)] void ondatasetchanged(IHTMLEventObj evtObj);
-            [DispId(-2147418097)] void ondataavailable(IHTMLEventObj evtObj);
-            [DispId(-2147418096)] void ondatasetcomplete(IHTMLEventObj evtObj);
-            [DispId(-2147418094)] void onlosecapture(IHTMLEventObj evtObj);
-            [DispId(-2147418093)] void onpropertychange(IHTMLEventObj evtObj);
-            [DispId(1014)] void onscroll(IHTMLEventObj evtObj);
-            [DispId(-2147418111)] void onfocus(IHTMLEventObj evtObj);
-            [DispId(-2147418112)] void onblur(IHTMLEventObj evtObj);
-            [DispId(1016)] void onresize(IHTMLEventObj evtObj);
-            [DispId(-2147418092)] bool ondrag(IHTMLEventObj evtObj);
-            [DispId(-2147418091)] void ondragend(IHTMLEventObj evtObj);
-            [DispId(-2147418090)] bool ondragenter(IHTMLEventObj evtObj);
-            [DispId(-2147418089)] bool ondragover(IHTMLEventObj evtObj);
-            [DispId(-2147418088)] void ondragleave(IHTMLEventObj evtObj);
-            [DispId(-2147418087)] bool ondrop(IHTMLEventObj evtObj);
-            [DispId(-2147418083)] bool onbeforecut(IHTMLEventObj evtObj);
-            [DispId(-2147418086)] bool oncut(IHTMLEventObj evtObj);
-            [DispId(-2147418082)] bool onbeforecopy(IHTMLEventObj evtObj);
-            [DispId(-2147418085)] bool oncopy(IHTMLEventObj evtObj);
-            [DispId(-2147418081)] bool onbeforepaste(IHTMLEventObj evtObj);
-            [DispId(-2147418084)] bool onpaste(IHTMLEventObj evtObj);
-            [DispId(1023)] bool oncontextmenu(IHTMLEventObj evtObj);
-            [DispId(-2147418080)] void onrowsdelete(IHTMLEventObj evtObj);
-            [DispId(-2147418079)] void onrowsinserted(IHTMLEventObj evtObj);
-            [DispId(-2147418078)] void oncellchange(IHTMLEventObj evtObj);
-            [DispId(-609)] void onreadystatechange(IHTMLEventObj evtObj);
-            [DispId(1030)] void onlayoutcomplete(IHTMLEventObj evtObj);
-            [DispId(1031)] void onpage(IHTMLEventObj evtObj);
-            [DispId(1042)] void onmouseenter(IHTMLEventObj evtObj);
-            [DispId(1043)] void onmouseleave(IHTMLEventObj evtObj);
-            [DispId(1044)] void onactivate(IHTMLEventObj evtObj);
-            [DispId(1045)] void ondeactivate(IHTMLEventObj evtObj);
-            [DispId(1034)] bool onbeforedeactivate(IHTMLEventObj evtObj);
-            [DispId(1047)] bool onbeforeactivate(IHTMLEventObj evtObj);
-            [DispId(1048)] void onfocusin(IHTMLEventObj evtObj);
-            [DispId(1049)] void onfocusout(IHTMLEventObj evtObj);
-            [DispId(1035)] void onmove(IHTMLEventObj evtObj);
-            [DispId(1036)] bool oncontrolselect(IHTMLEventObj evtObj);
-            [DispId(1038)] bool onmovestart(IHTMLEventObj evtObj);
-            [DispId(1039)] void onmoveend(IHTMLEventObj evtObj);
-            [DispId(1040)] bool onresizestart(IHTMLEventObj evtObj);
-            [DispId(1041)] void onresizeend(IHTMLEventObj evtObj);
-            [DispId(1033)] bool onmousewheel(IHTMLEventObj evtObj);
-            [DispId(-2147412092)] void onbounce(IHTMLEventObj evtObj);
-            [DispId(-2147412086)] void onfinish(IHTMLEventObj evtObj);
-            [DispId(-2147412085)] void onstart(IHTMLEventObj evtObj);
-        }
-
-        [ComImport(), Guid("3050f619-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsIDispatch),
-        TypeLibType(TypeLibTypeFlags.FHidden)]
-        public interface DHTMLOptionButtonElementEvents2
-        {
-            [DispId(-2147418102)] bool onhelp(IHTMLEventObj evtObj);
-            [DispId(-600)] bool onclick(IHTMLEventObj evtObj);
-            [DispId(-601)] bool ondblclick(IHTMLEventObj evtObj);
-            [DispId(-603)] bool onkeypress(IHTMLEventObj evtObj);
-            [DispId(-602)] void onkeydown(IHTMLEventObj evtObj);
-            [DispId(-604)] void onkeyup(IHTMLEventObj evtObj);
-            [DispId(-2147418103)] void onmouseout(IHTMLEventObj evtObj);
-            [DispId(-2147418104)] void onmouseover(IHTMLEventObj evtObj);
-            [DispId(-606)] void onmousemove(IHTMLEventObj evtObj);
-            [DispId(-605)] void onmousedown(IHTMLEventObj evtObj);
-            [DispId(-607)] void onmouseup(IHTMLEventObj evtObj);
-            [DispId(-2147418100)] bool onselectstart(IHTMLEventObj evtObj);
-            [DispId(-2147418095)] void onfilterchange(IHTMLEventObj evtObj);
-            [DispId(-2147418101)] bool ondragstart(IHTMLEventObj evtObj);
-            [DispId(-2147418108)] bool onbeforeupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418107)] void onafterupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418099)] bool onerrorupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418106)] bool onrowexit(IHTMLEventObj evtObj);
-            [DispId(-2147418105)] void onrowenter(IHTMLEventObj evtObj);
-            [DispId(-2147418098)] void ondatasetchanged(IHTMLEventObj evtObj);
-            [DispId(-2147418097)] void ondataavailable(IHTMLEventObj evtObj);
-            [DispId(-2147418096)] void ondatasetcomplete(IHTMLEventObj evtObj);
-            [DispId(-2147418094)] void onlosecapture(IHTMLEventObj evtObj);
-            [DispId(-2147418093)] void onpropertychange(IHTMLEventObj evtObj);
-            [DispId(1014)] void onscroll(IHTMLEventObj evtObj);
-            [DispId(-2147418111)] void onfocus(IHTMLEventObj evtObj);
-            [DispId(-2147418112)] void onblur(IHTMLEventObj evtObj);
-            [DispId(1016)] void onresize(IHTMLEventObj evtObj);
-            [DispId(-2147418092)] bool ondrag(IHTMLEventObj evtObj);
-            [DispId(-2147418091)] void ondragend(IHTMLEventObj evtObj);
-            [DispId(-2147418090)] bool ondragenter(IHTMLEventObj evtObj);
-            [DispId(-2147418089)] bool ondragover(IHTMLEventObj evtObj);
-            [DispId(-2147418088)] void ondragleave(IHTMLEventObj evtObj);
-            [DispId(-2147418087)] bool ondrop(IHTMLEventObj evtObj);
-            [DispId(-2147418083)] bool onbeforecut(IHTMLEventObj evtObj);
-            [DispId(-2147418086)] bool oncut(IHTMLEventObj evtObj);
-            [DispId(-2147418082)] bool onbeforecopy(IHTMLEventObj evtObj);
-            [DispId(-2147418085)] bool oncopy(IHTMLEventObj evtObj);
-            [DispId(-2147418081)] bool onbeforepaste(IHTMLEventObj evtObj);
-            [DispId(-2147418084)] bool onpaste(IHTMLEventObj evtObj);
-            [DispId(1023)] bool oncontextmenu(IHTMLEventObj evtObj);
-            [DispId(-2147418080)] void onrowsdelete(IHTMLEventObj evtObj);
-            [DispId(-2147418079)] void onrowsinserted(IHTMLEventObj evtObj);
-            [DispId(-2147418078)] void oncellchange(IHTMLEventObj evtObj);
-            [DispId(-609)] void onreadystatechange(IHTMLEventObj evtObj);
-            [DispId(1030)] void onlayoutcomplete(IHTMLEventObj evtObj);
-            [DispId(1031)] void onpage(IHTMLEventObj evtObj);
-            [DispId(1042)] void onmouseenter(IHTMLEventObj evtObj);
-            [DispId(1043)] void onmouseleave(IHTMLEventObj evtObj);
-            [DispId(1044)] void onactivate(IHTMLEventObj evtObj);
-            [DispId(1045)] void ondeactivate(IHTMLEventObj evtObj);
-            [DispId(1034)] bool onbeforedeactivate(IHTMLEventObj evtObj);
-            [DispId(1047)] bool onbeforeactivate(IHTMLEventObj evtObj);
-            [DispId(1048)] void onfocusin(IHTMLEventObj evtObj);
-            [DispId(1049)] void onfocusout(IHTMLEventObj evtObj);
-            [DispId(1035)] void onmove(IHTMLEventObj evtObj);
-            [DispId(1036)] bool oncontrolselect(IHTMLEventObj evtObj);
-            [DispId(1038)] bool onmovestart(IHTMLEventObj evtObj);
-            [DispId(1039)] void onmoveend(IHTMLEventObj evtObj);
-            [DispId(1040)] bool onresizestart(IHTMLEventObj evtObj);
-            [DispId(1041)] void onresizeend(IHTMLEventObj evtObj);
-            [DispId(1033)] bool onmousewheel(IHTMLEventObj evtObj);
-            [DispId(-2147412082)] bool onchange(IHTMLEventObj evtObj);
-        }
-
-        [ComImport(), Guid("3050f622-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsIDispatch),
-        TypeLibType(TypeLibTypeFlags.FHidden)]
-        public interface DHTMLSelectElementEvents2
-        {
-            [DispId(-2147418102)] bool onhelp(IHTMLEventObj evtObj);
-            [DispId(-600)] bool onclick(IHTMLEventObj evtObj);
-            [DispId(-601)] bool ondblclick(IHTMLEventObj evtObj);
-            [DispId(-603)] bool onkeypress(IHTMLEventObj evtObj);
-            [DispId(-602)] void onkeydown(IHTMLEventObj evtObj);
-            [DispId(-604)] void onkeyup(IHTMLEventObj evtObj);
-            [DispId(-2147418103)] void onmouseout(IHTMLEventObj evtObj);
-            [DispId(-2147418104)] void onmouseover(IHTMLEventObj evtObj);
-            [DispId(-606)] void onmousemove(IHTMLEventObj evtObj);
-            [DispId(-605)] void onmousedown(IHTMLEventObj evtObj);
-            [DispId(-607)] void onmouseup(IHTMLEventObj evtObj);
-            [DispId(-2147418100)] bool onselectstart(IHTMLEventObj evtObj);
-            [DispId(-2147418095)] void onfilterchange(IHTMLEventObj evtObj);
-            [DispId(-2147418101)] bool ondragstart(IHTMLEventObj evtObj);
-            [DispId(-2147418108)] bool onbeforeupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418107)] void onafterupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418099)] bool onerrorupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418106)] bool onrowexit(IHTMLEventObj evtObj);
-            [DispId(-2147418105)] void onrowenter(IHTMLEventObj evtObj);
-            [DispId(-2147418098)] void ondatasetchanged(IHTMLEventObj evtObj);
-            [DispId(-2147418097)] void ondataavailable(IHTMLEventObj evtObj);
-            [DispId(-2147418096)] void ondatasetcomplete(IHTMLEventObj evtObj);
-            [DispId(-2147418094)] void onlosecapture(IHTMLEventObj evtObj);
-            [DispId(-2147418093)] void onpropertychange(IHTMLEventObj evtObj);
-            [DispId(1014)] void onscroll(IHTMLEventObj evtObj);
-            [DispId(-2147418111)] void onfocus(IHTMLEventObj evtObj);
-            [DispId(-2147418112)] void onblur(IHTMLEventObj evtObj);
-            [DispId(1016)] void onresize(IHTMLEventObj evtObj);
-            [DispId(-2147418092)] bool ondrag(IHTMLEventObj evtObj);
-            [DispId(-2147418091)] void ondragend(IHTMLEventObj evtObj);
-            [DispId(-2147418090)] bool ondragenter(IHTMLEventObj evtObj);
-            [DispId(-2147418089)] bool ondragover(IHTMLEventObj evtObj);
-            [DispId(-2147418088)] void ondragleave(IHTMLEventObj evtObj);
-            [DispId(-2147418087)] bool ondrop(IHTMLEventObj evtObj);
-            [DispId(-2147418083)] bool onbeforecut(IHTMLEventObj evtObj);
-            [DispId(-2147418086)] bool oncut(IHTMLEventObj evtObj);
-            [DispId(-2147418082)] bool onbeforecopy(IHTMLEventObj evtObj);
-            [DispId(-2147418085)] bool oncopy(IHTMLEventObj evtObj);
-            [DispId(-2147418081)] bool onbeforepaste(IHTMLEventObj evtObj);
-            [DispId(-2147418084)] bool onpaste(IHTMLEventObj evtObj);
-            [DispId(1023)] bool oncontextmenu(IHTMLEventObj evtObj);
-            [DispId(-2147418080)] void onrowsdelete(IHTMLEventObj evtObj);
-            [DispId(-2147418079)] void onrowsinserted(IHTMLEventObj evtObj);
-            [DispId(-2147418078)] void oncellchange(IHTMLEventObj evtObj);
-            [DispId(-609)] void onreadystatechange(IHTMLEventObj evtObj);
-            [DispId(1030)] void onlayoutcomplete(IHTMLEventObj evtObj);
-            [DispId(1031)] void onpage(IHTMLEventObj evtObj);
-            [DispId(1042)] void onmouseenter(IHTMLEventObj evtObj);
-            [DispId(1043)] void onmouseleave(IHTMLEventObj evtObj);
-            [DispId(1044)] void onactivate(IHTMLEventObj evtObj);
-            [DispId(1045)] void ondeactivate(IHTMLEventObj evtObj);
-            [DispId(1034)] bool onbeforedeactivate(IHTMLEventObj evtObj);
-            [DispId(1047)] bool onbeforeactivate(IHTMLEventObj evtObj);
-            [DispId(1048)] void onfocusin(IHTMLEventObj evtObj);
-            [DispId(1049)] void onfocusout(IHTMLEventObj evtObj);
-            [DispId(1035)] void onmove(IHTMLEventObj evtObj);
-            [DispId(1036)] bool oncontrolselect(IHTMLEventObj evtObj);
-            [DispId(1038)] bool onmovestart(IHTMLEventObj evtObj);
-            [DispId(1039)] void onmoveend(IHTMLEventObj evtObj);
-            [DispId(1040)] bool onresizestart(IHTMLEventObj evtObj);
-            [DispId(1041)] void onresizeend(IHTMLEventObj evtObj);
-            [DispId(1033)] bool onmousewheel(IHTMLEventObj evtObj);
-            [DispId(-2147412082)] void onchange_void(IHTMLEventObj evtObj);
-        }
-
-        [ComImport(), Guid("3050f615-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsIDispatch),
-        TypeLibType(TypeLibTypeFlags.FHidden)]
-        public interface DHTMLStyleElementEvents2
-        {
-            [DispId(-2147418102)] bool onhelp(IHTMLEventObj evtObj);
-            [DispId(-600)] bool onclick(IHTMLEventObj evtObj);
-            [DispId(-601)] bool ondblclick(IHTMLEventObj evtObj);
-            [DispId(-603)] bool onkeypress(IHTMLEventObj evtObj);
-            [DispId(-602)] void onkeydown(IHTMLEventObj evtObj);
-            [DispId(-604)] void onkeyup(IHTMLEventObj evtObj);
-            [DispId(-2147418103)] void onmouseout(IHTMLEventObj evtObj);
-            [DispId(-2147418104)] void onmouseover(IHTMLEventObj evtObj);
-            [DispId(-606)] void onmousemove(IHTMLEventObj evtObj);
-            [DispId(-605)] void onmousedown(IHTMLEventObj evtObj);
-            [DispId(-607)] void onmouseup(IHTMLEventObj evtObj);
-            [DispId(-2147418100)] bool onselectstart(IHTMLEventObj evtObj);
-            [DispId(-2147418095)] void onfilterchange(IHTMLEventObj evtObj);
-            [DispId(-2147418101)] bool ondragstart(IHTMLEventObj evtObj);
-            [DispId(-2147418108)] bool onbeforeupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418107)] void onafterupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418099)] bool onerrorupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418106)] bool onrowexit(IHTMLEventObj evtObj);
-            [DispId(-2147418105)] void onrowenter(IHTMLEventObj evtObj);
-            [DispId(-2147418098)] void ondatasetchanged(IHTMLEventObj evtObj);
-            [DispId(-2147418097)] void ondataavailable(IHTMLEventObj evtObj);
-            [DispId(-2147418096)] void ondatasetcomplete(IHTMLEventObj evtObj);
-            [DispId(-2147418094)] void onlosecapture(IHTMLEventObj evtObj);
-            [DispId(-2147418093)] void onpropertychange(IHTMLEventObj evtObj);
-            [DispId(1014)] void onscroll(IHTMLEventObj evtObj);
-            [DispId(-2147418111)] void onfocus(IHTMLEventObj evtObj);
-            [DispId(-2147418112)] void onblur(IHTMLEventObj evtObj);
-            [DispId(1016)] void onresize(IHTMLEventObj evtObj);
-            [DispId(-2147418092)] bool ondrag(IHTMLEventObj evtObj);
-            [DispId(-2147418091)] void ondragend(IHTMLEventObj evtObj);
-            [DispId(-2147418090)] bool ondragenter(IHTMLEventObj evtObj);
-            [DispId(-2147418089)] bool ondragover(IHTMLEventObj evtObj);
-            [DispId(-2147418088)] void ondragleave(IHTMLEventObj evtObj);
-            [DispId(-2147418087)] bool ondrop(IHTMLEventObj evtObj);
-            [DispId(-2147418083)] bool onbeforecut(IHTMLEventObj evtObj);
-            [DispId(-2147418086)] bool oncut(IHTMLEventObj evtObj);
-            [DispId(-2147418082)] bool onbeforecopy(IHTMLEventObj evtObj);
-            [DispId(-2147418085)] bool oncopy(IHTMLEventObj evtObj);
-            [DispId(-2147418081)] bool onbeforepaste(IHTMLEventObj evtObj);
-            [DispId(-2147418084)] bool onpaste(IHTMLEventObj evtObj);
-            [DispId(1023)] bool oncontextmenu(IHTMLEventObj evtObj);
-            [DispId(-2147418080)] void onrowsdelete(IHTMLEventObj evtObj);
-            [DispId(-2147418079)] void onrowsinserted(IHTMLEventObj evtObj);
-            [DispId(-2147418078)] void oncellchange(IHTMLEventObj evtObj);
-            [DispId(-609)] void onreadystatechange(IHTMLEventObj evtObj);
-            [DispId(1030)] void onlayoutcomplete(IHTMLEventObj evtObj);
-            [DispId(1031)] void onpage(IHTMLEventObj evtObj);
-            [DispId(1042)] void onmouseenter(IHTMLEventObj evtObj);
-            [DispId(1043)] void onmouseleave(IHTMLEventObj evtObj);
-            [DispId(1044)] void onactivate(IHTMLEventObj evtObj);
-            [DispId(1045)] void ondeactivate(IHTMLEventObj evtObj);
-            [DispId(1034)] bool onbeforedeactivate(IHTMLEventObj evtObj);
-            [DispId(1047)] bool onbeforeactivate(IHTMLEventObj evtObj);
-            [DispId(1048)] void onfocusin(IHTMLEventObj evtObj);
-            [DispId(1049)] void onfocusout(IHTMLEventObj evtObj);
-            [DispId(1035)] void onmove(IHTMLEventObj evtObj);
-            [DispId(1036)] bool oncontrolselect(IHTMLEventObj evtObj);
-            [DispId(1038)] bool onmovestart(IHTMLEventObj evtObj);
-            [DispId(1039)] void onmoveend(IHTMLEventObj evtObj);
-            [DispId(1040)] bool onresizestart(IHTMLEventObj evtObj);
-            [DispId(1041)] void onresizeend(IHTMLEventObj evtObj);
-            [DispId(1033)] bool onmousewheel(IHTMLEventObj evtObj);
-            [DispId(1003)] void onload(IHTMLEventObj evtObj);
-            [DispId(1002)] void onerror(IHTMLEventObj evtObj);
-        }
-
-        [ComImport(), Guid("3050f623-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsIDispatch),
-        TypeLibType(TypeLibTypeFlags.FHidden)]
-        public interface DHTMLTableEvents2
-        {
-            [DispId(-2147418102)] bool onhelp(IHTMLEventObj evtObj);
-            [DispId(-600)] bool onclick(IHTMLEventObj evtObj);
-            [DispId(-601)] bool ondblclick(IHTMLEventObj evtObj);
-            [DispId(-603)] bool onkeypress(IHTMLEventObj evtObj);
-            [DispId(-602)] void onkeydown(IHTMLEventObj evtObj);
-            [DispId(-604)] void onkeyup(IHTMLEventObj evtObj);
-            [DispId(-2147418103)] void onmouseout(IHTMLEventObj evtObj);
-            [DispId(-2147418104)] void onmouseover(IHTMLEventObj evtObj);
-            [DispId(-606)] void onmousemove(IHTMLEventObj evtObj);
-            [DispId(-605)] void onmousedown(IHTMLEventObj evtObj);
-            [DispId(-607)] void onmouseup(IHTMLEventObj evtObj);
-            [DispId(-2147418100)] bool onselectstart(IHTMLEventObj evtObj);
-            [DispId(-2147418095)] void onfilterchange(IHTMLEventObj evtObj);
-            [DispId(-2147418101)] bool ondragstart(IHTMLEventObj evtObj);
-            [DispId(-2147418108)] bool onbeforeupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418107)] void onafterupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418099)] bool onerrorupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418106)] bool onrowexit(IHTMLEventObj evtObj);
-            [DispId(-2147418105)] void onrowenter(IHTMLEventObj evtObj);
-            [DispId(-2147418098)] void ondatasetchanged(IHTMLEventObj evtObj);
-            [DispId(-2147418097)] void ondataavailable(IHTMLEventObj evtObj);
-            [DispId(-2147418096)] void ondatasetcomplete(IHTMLEventObj evtObj);
-            [DispId(-2147418094)] void onlosecapture(IHTMLEventObj evtObj);
-            [DispId(-2147418093)] void onpropertychange(IHTMLEventObj evtObj);
-            [DispId(1014)] void onscroll(IHTMLEventObj evtObj);
-            [DispId(-2147418111)] void onfocus(IHTMLEventObj evtObj);
-            [DispId(-2147418112)] void onblur(IHTMLEventObj evtObj);
-            [DispId(1016)] void onresize(IHTMLEventObj evtObj);
-            [DispId(-2147418092)] bool ondrag(IHTMLEventObj evtObj);
-            [DispId(-2147418091)] void ondragend(IHTMLEventObj evtObj);
-            [DispId(-2147418090)] bool ondragenter(IHTMLEventObj evtObj);
-            [DispId(-2147418089)] bool ondragover(IHTMLEventObj evtObj);
-            [DispId(-2147418088)] void ondragleave(IHTMLEventObj evtObj);
-            [DispId(-2147418087)] bool ondrop(IHTMLEventObj evtObj);
-            [DispId(-2147418083)] bool onbeforecut(IHTMLEventObj evtObj);
-            [DispId(-2147418086)] bool oncut(IHTMLEventObj evtObj);
-            [DispId(-2147418082)] bool onbeforecopy(IHTMLEventObj evtObj);
-            [DispId(-2147418085)] bool oncopy(IHTMLEventObj evtObj);
-            [DispId(-2147418081)] bool onbeforepaste(IHTMLEventObj evtObj);
-            [DispId(-2147418084)] bool onpaste(IHTMLEventObj evtObj);
-            [DispId(1023)] bool oncontextmenu(IHTMLEventObj evtObj);
-            [DispId(-2147418080)] void onrowsdelete(IHTMLEventObj evtObj);
-            [DispId(-2147418079)] void onrowsinserted(IHTMLEventObj evtObj);
-            [DispId(-2147418078)] void oncellchange(IHTMLEventObj evtObj);
-            [DispId(-609)] void onreadystatechange(IHTMLEventObj evtObj);
-            [DispId(1030)] void onlayoutcomplete(IHTMLEventObj evtObj);
-            [DispId(1031)] void onpage(IHTMLEventObj evtObj);
-            [DispId(1042)] void onmouseenter(IHTMLEventObj evtObj);
-            [DispId(1043)] void onmouseleave(IHTMLEventObj evtObj);
-            [DispId(1044)] void onactivate(IHTMLEventObj evtObj);
-            [DispId(1045)] void ondeactivate(IHTMLEventObj evtObj);
-            [DispId(1034)] bool onbeforedeactivate(IHTMLEventObj evtObj);
-            [DispId(1047)] bool onbeforeactivate(IHTMLEventObj evtObj);
-            [DispId(1048)] void onfocusin(IHTMLEventObj evtObj);
-            [DispId(1049)] void onfocusout(IHTMLEventObj evtObj);
-            [DispId(1035)] void onmove(IHTMLEventObj evtObj);
-            [DispId(1036)] bool oncontrolselect(IHTMLEventObj evtObj);
-            [DispId(1038)] bool onmovestart(IHTMLEventObj evtObj);
-            [DispId(1039)] void onmoveend(IHTMLEventObj evtObj);
-            [DispId(1040)] bool onresizestart(IHTMLEventObj evtObj);
-            [DispId(1041)] void onresizeend(IHTMLEventObj evtObj);
-            [DispId(1033)] bool onmousewheel(IHTMLEventObj evtObj);
-        }
-
-        [ComImport(), Guid("3050f624-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsIDispatch),
-        TypeLibType(TypeLibTypeFlags.FHidden)]
-        public interface DHTMLTextContainerEvents2
-        {
-            [DispId(-2147418102)] bool onhelp(IHTMLEventObj evtObj);
-            [DispId(-600)] bool onclick(IHTMLEventObj evtObj);
-            [DispId(-601)] bool ondblclick(IHTMLEventObj evtObj);
-            [DispId(-603)] bool onkeypress(IHTMLEventObj evtObj);
-            [DispId(-602)] void onkeydown(IHTMLEventObj evtObj);
-            [DispId(-604)] void onkeyup(IHTMLEventObj evtObj);
-            [DispId(-2147418103)] void onmouseout(IHTMLEventObj evtObj);
-            [DispId(-2147418104)] void onmouseover(IHTMLEventObj evtObj);
-            [DispId(-606)] void onmousemove(IHTMLEventObj evtObj);
-            [DispId(-605)] void onmousedown(IHTMLEventObj evtObj);
-            [DispId(-607)] void onmouseup(IHTMLEventObj evtObj);
-            [DispId(-2147418100)] bool onselectstart(IHTMLEventObj evtObj);
-            [DispId(-2147418095)] void onfilterchange(IHTMLEventObj evtObj);
-            [DispId(-2147418101)] bool ondragstart(IHTMLEventObj evtObj);
-            [DispId(-2147418108)] bool onbeforeupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418107)] void onafterupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418099)] bool onerrorupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418106)] bool onrowexit(IHTMLEventObj evtObj);
-            [DispId(-2147418105)] void onrowenter(IHTMLEventObj evtObj);
-            [DispId(-2147418098)] void ondatasetchanged(IHTMLEventObj evtObj);
-            [DispId(-2147418097)] void ondataavailable(IHTMLEventObj evtObj);
-            [DispId(-2147418096)] void ondatasetcomplete(IHTMLEventObj evtObj);
-            [DispId(-2147418094)] void onlosecapture(IHTMLEventObj evtObj);
-            [DispId(-2147418093)] void onpropertychange(IHTMLEventObj evtObj);
-            [DispId(1014)] void onscroll(IHTMLEventObj evtObj);
-            [DispId(-2147418111)] void onfocus(IHTMLEventObj evtObj);
-            [DispId(-2147418112)] void onblur(IHTMLEventObj evtObj);
-            [DispId(1016)] void onresize(IHTMLEventObj evtObj);
-            [DispId(-2147418092)] bool ondrag(IHTMLEventObj evtObj);
-            [DispId(-2147418091)] void ondragend(IHTMLEventObj evtObj);
-            [DispId(-2147418090)] bool ondragenter(IHTMLEventObj evtObj);
-            [DispId(-2147418089)] bool ondragover(IHTMLEventObj evtObj);
-            [DispId(-2147418088)] void ondragleave(IHTMLEventObj evtObj);
-            [DispId(-2147418087)] bool ondrop(IHTMLEventObj evtObj);
-            [DispId(-2147418083)] bool onbeforecut(IHTMLEventObj evtObj);
-            [DispId(-2147418086)] bool oncut(IHTMLEventObj evtObj);
-            [DispId(-2147418082)] bool onbeforecopy(IHTMLEventObj evtObj);
-            [DispId(-2147418085)] bool oncopy(IHTMLEventObj evtObj);
-            [DispId(-2147418081)] bool onbeforepaste(IHTMLEventObj evtObj);
-            [DispId(-2147418084)] bool onpaste(IHTMLEventObj evtObj);
-            [DispId(1023)] bool oncontextmenu(IHTMLEventObj evtObj);
-            [DispId(-2147418080)] void onrowsdelete(IHTMLEventObj evtObj);
-            [DispId(-2147418079)] void onrowsinserted(IHTMLEventObj evtObj);
-            [DispId(-2147418078)] void oncellchange(IHTMLEventObj evtObj);
-            [DispId(-609)] void onreadystatechange(IHTMLEventObj evtObj);
-            [DispId(1030)] void onlayoutcomplete(IHTMLEventObj evtObj);
-            [DispId(1031)] void onpage(IHTMLEventObj evtObj);
-            [DispId(1042)] void onmouseenter(IHTMLEventObj evtObj);
-            [DispId(1043)] void onmouseleave(IHTMLEventObj evtObj);
-            [DispId(1044)] void onactivate(IHTMLEventObj evtObj);
-            [DispId(1045)] void ondeactivate(IHTMLEventObj evtObj);
-            [DispId(1034)] bool onbeforedeactivate(IHTMLEventObj evtObj);
-            [DispId(1047)] bool onbeforeactivate(IHTMLEventObj evtObj);
-            [DispId(1048)] void onfocusin(IHTMLEventObj evtObj);
-            [DispId(1049)] void onfocusout(IHTMLEventObj evtObj);
-            [DispId(1035)] void onmove(IHTMLEventObj evtObj);
-            [DispId(1036)] bool oncontrolselect(IHTMLEventObj evtObj);
-            [DispId(1038)] bool onmovestart(IHTMLEventObj evtObj);
-            [DispId(1039)] void onmoveend(IHTMLEventObj evtObj);
-            [DispId(1040)] bool onresizestart(IHTMLEventObj evtObj);
-            [DispId(1041)] void onresizeend(IHTMLEventObj evtObj);
-            [DispId(1033)] bool onmousewheel(IHTMLEventObj evtObj);
-            [DispId(1001)] void onchange_void(IHTMLEventObj evtObj);
-            [DispId(1006)] void onselect(IHTMLEventObj evtObj);
-        }
-
-        [ComImport(), Guid("3050f621-98b5-11cf-bb82-00aa00bdce0b"),
-        InterfaceType(ComInterfaceType.InterfaceIsIDispatch),
-        TypeLibType(TypeLibTypeFlags.FHidden)]
-        public interface DHTMLScriptEvents2
-        {
-            [DispId(-2147418102)] bool onhelp(IHTMLEventObj evtObj);
-            [DispId(-600)] bool onclick(IHTMLEventObj evtObj);
-            [DispId(-601)] bool ondblclick(IHTMLEventObj evtObj);
-            [DispId(-603)] bool onkeypress(IHTMLEventObj evtObj);
-            [DispId(-602)] void onkeydown(IHTMLEventObj evtObj);
-            [DispId(-604)] void onkeyup(IHTMLEventObj evtObj);
-            [DispId(-2147418103)] void onmouseout(IHTMLEventObj evtObj);
-            [DispId(-2147418104)] void onmouseover(IHTMLEventObj evtObj);
-            [DispId(-606)] void onmousemove(IHTMLEventObj evtObj);
-            [DispId(-605)] void onmousedown(IHTMLEventObj evtObj);
-            [DispId(-607)] void onmouseup(IHTMLEventObj evtObj);
-            [DispId(-2147418100)] bool onselectstart(IHTMLEventObj evtObj);
-            [DispId(-2147418095)] void onfilterchange(IHTMLEventObj evtObj);
-            [DispId(-2147418101)] bool ondragstart(IHTMLEventObj evtObj);
-            [DispId(-2147418108)] bool onbeforeupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418107)] void onafterupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418099)] bool onerrorupdate(IHTMLEventObj evtObj);
-            [DispId(-2147418106)] bool onrowexit(IHTMLEventObj evtObj);
-            [DispId(-2147418105)] void onrowenter(IHTMLEventObj evtObj);
-            [DispId(-2147418098)] void ondatasetchanged(IHTMLEventObj evtObj);
-            [DispId(-2147418097)] void ondataavailable(IHTMLEventObj evtObj);
-            [DispId(-2147418096)] void ondatasetcomplete(IHTMLEventObj evtObj);
-            [DispId(-2147418094)] void onlosecapture(IHTMLEventObj evtObj);
-            [DispId(-2147418093)] void onpropertychange(IHTMLEventObj evtObj);
-            [DispId(1014)] void onscroll(IHTMLEventObj evtObj);
-            [DispId(-2147418111)] void onfocus(IHTMLEventObj evtObj);
-            [DispId(-2147418112)] void onblur(IHTMLEventObj evtObj);
-            [DispId(1016)] void onresize(IHTMLEventObj evtObj);
-            [DispId(-2147418092)] bool ondrag(IHTMLEventObj evtObj);
-            [DispId(-2147418091)] void ondragend(IHTMLEventObj evtObj);
-            [DispId(-2147418090)] bool ondragenter(IHTMLEventObj evtObj);
-            [DispId(-2147418089)] bool ondragover(IHTMLEventObj evtObj);
-            [DispId(-2147418088)] void ondragleave(IHTMLEventObj evtObj);
-            [DispId(-2147418087)] bool ondrop(IHTMLEventObj evtObj);
-            [DispId(-2147418083)] bool onbeforecut(IHTMLEventObj evtObj);
-            [DispId(-2147418086)] bool oncut(IHTMLEventObj evtObj);
-            [DispId(-2147418082)] bool onbeforecopy(IHTMLEventObj evtObj);
-            [DispId(-2147418085)] bool oncopy(IHTMLEventObj evtObj);
-            [DispId(-2147418081)] bool onbeforepaste(IHTMLEventObj evtObj);
-            [DispId(-2147418084)] bool onpaste(IHTMLEventObj evtObj);
-            [DispId(1023)] bool oncontextmenu(IHTMLEventObj evtObj);
-            [DispId(-2147418080)] void onrowsdelete(IHTMLEventObj evtObj);
-            [DispId(-2147418079)] void onrowsinserted(IHTMLEventObj evtObj);
-            [DispId(-2147418078)] void oncellchange(IHTMLEventObj evtObj);
-            [DispId(-609)] void onreadystatechange(IHTMLEventObj evtObj);
-            [DispId(1030)] void onlayoutcomplete(IHTMLEventObj evtObj);
-            [DispId(1031)] void onpage(IHTMLEventObj evtObj);
-            [DispId(1042)] void onmouseenter(IHTMLEventObj evtObj);
-            [DispId(1043)] void onmouseleave(IHTMLEventObj evtObj);
-            [DispId(1044)] void onactivate(IHTMLEventObj evtObj);
-            [DispId(1045)] void ondeactivate(IHTMLEventObj evtObj);
-            [DispId(1034)] bool onbeforedeactivate(IHTMLEventObj evtObj);
-            [DispId(1047)] bool onbeforeactivate(IHTMLEventObj evtObj);
-            [DispId(1048)] void onfocusin(IHTMLEventObj evtObj);
-            [DispId(1049)] void onfocusout(IHTMLEventObj evtObj);
-            [DispId(1035)] void onmove(IHTMLEventObj evtObj);
-            [DispId(1036)] bool oncontrolselect(IHTMLEventObj evtObj);
-            [DispId(1038)] bool onmovestart(IHTMLEventObj evtObj);
-            [DispId(1039)] void onmoveend(IHTMLEventObj evtObj);
-            [DispId(1040)] bool onresizestart(IHTMLEventObj evtObj);
-            [DispId(1041)] void onresizeend(IHTMLEventObj evtObj);
-            [DispId(1033)] bool onmousewheel(IHTMLEventObj evtObj);
-            [DispId(1002)] void onerror(IHTMLEventObj evtObj);
-        }
-
-        [ComVisible(true), Guid("3050F25E-98B5-11CF-BB82-00AA00BDCE0B"),
-        InterfaceType(ComInterfaceType.InterfaceIsDual)]
-        internal interface IHTMLStyle
-        {
-            void SetFontFamily(string p);
-            string GetFontFamily();
-            void SetFontStyle(string p);
-            string GetFontStyle();
-            void SetFontObject(string p);
-            string GetFontObject();
-            void SetFontWeight(string p);
-            string GetFontWeight();
-            void SetFontSize(object p);
-            object GetFontSize();
-            void SetFont(string p);
-            string GetFont();
-            void SetColor(object p);
-            object GetColor();
-            void SetBackground(string p);
-            string GetBackground();
-            void SetBackgroundColor(object p);
-            object GetBackgroundColor();
-            void SetBackgroundImage(string p);
-            string GetBackgroundImage();
-            void SetBackgroundRepeat(string p);
-            string GetBackgroundRepeat();
-            void SetBackgroundAttachment(string p);
-            string GetBackgroundAttachment();
-            void SetBackgroundPosition(string p);
-            string GetBackgroundPosition();
-            void SetBackgroundPositionX(object p);
-            object GetBackgroundPositionX();
-            void SetBackgroundPositionY(object p);
-            object GetBackgroundPositionY();
-            void SetWordSpacing(object p);
-            object GetWordSpacing();
-            void SetLetterSpacing(object p);
-            object GetLetterSpacing();
-            void SetTextDecoration(string p);
-            string GetTextDecoration();
-            void SetTextDecorationNone(bool p);
-            bool GetTextDecorationNone();
-            void SetTextDecorationUnderline(bool p);
-            bool GetTextDecorationUnderline();
-            void SetTextDecorationOverline(bool p);
-            bool GetTextDecorationOverline();
-            void SetTextDecorationLineThrough(bool p);
-            bool GetTextDecorationLineThrough();
-            void SetTextDecorationBlink(bool p);
-            bool GetTextDecorationBlink();
-            void SetVerticalAlign(object p);
-            object GetVerticalAlign();
-            void SetTextTransform(string p);
-            string GetTextTransform();
-            void SetTextAlign(string p);
-            string GetTextAlign();
-            void SetTextIndent(object p);
-            object GetTextIndent();
-            void SetLineHeight(object p);
-            object GetLineHeight();
-            void SetMarginTop(object p);
-            object GetMarginTop();
-            void SetMarginRight(object p);
-            object GetMarginRight();
-            void SetMarginBottom(object p);
-            object GetMarginBottom();
-            void SetMarginLeft(object p);
-            object GetMarginLeft();
-            void SetMargin(string p);
-            string GetMargin();
-            void SetPaddingTop(object p);
-            object GetPaddingTop();
-            void SetPaddingRight(object p);
-            object GetPaddingRight();
-            void SetPaddingBottom(object p);
-            object GetPaddingBottom();
-            void SetPaddingLeft(object p);
-            object GetPaddingLeft();
-            void SetPadding(string p);
-            string GetPadding();
-            void SetBorder(string p);
-            string GetBorder();
-            void SetBorderTop(string p);
-            string GetBorderTop();
-            void SetBorderRight(string p);
-            string GetBorderRight();
-            void SetBorderBottom(string p);
-            string GetBorderBottom();
-            void SetBorderLeft(string p);
-            string GetBorderLeft();
-            void SetBorderColor(string p);
-            string GetBorderColor();
-            void SetBorderTopColor(object p);
-            object GetBorderTopColor();
-            void SetBorderRightColor(object p);
-            object GetBorderRightColor();
-            void SetBorderBottomColor(object p);
-            object GetBorderBottomColor();
-            void SetBorderLeftColor(object p);
-            object GetBorderLeftColor();
-            void SetBorderWidth(string p);
-            string GetBorderWidth();
-            void SetBorderTopWidth(object p);
-            object GetBorderTopWidth();
-            void SetBorderRightWidth(object p);
-            object GetBorderRightWidth();
-            void SetBorderBottomWidth(object p);
-            object GetBorderBottomWidth();
-            void SetBorderLeftWidth(object p);
-            object GetBorderLeftWidth();
-            void SetBorderStyle(string p);
-            string GetBorderStyle();
-            void SetBorderTopStyle(string p);
-            string GetBorderTopStyle();
-            void SetBorderRightStyle(string p);
-            string GetBorderRightStyle();
-            void SetBorderBottomStyle(string p);
-            string GetBorderBottomStyle();
-            void SetBorderLeftStyle(string p);
-            string GetBorderLeftStyle();
-            void SetWidth(object p);
-            object GetWidth();
-            void SetHeight(object p);
-            object GetHeight();
-            void SetStyleFloat(string p);
-            string GetStyleFloat();
-            void SetClear(string p);
-            string GetClear();
-            void SetDisplay(string p);
-            string GetDisplay();
-            void SetVisibility(string p);
-            string GetVisibility();
-            void SetListStyleType(string p);
-            string GetListStyleType();
-            void SetListStylePosition(string p);
-            string GetListStylePosition();
-            void SetListStyleImage(string p);
-            string GetListStyleImage();
-            void SetListStyle(string p);
-            string GetListStyle();
-            void SetWhiteSpace(string p);
-            string GetWhiteSpace();
-            void SetTop(object p);
-            object GetTop();
-            void SetLeft(object p);
-            object GetLeft();
-            string GetPosition();
-            void SetZIndex(object p);
-            object GetZIndex();
-            void SetOverflow(string p);
-            string GetOverflow();
-            void SetPageBreakBefore(string p);
-            string GetPageBreakBefore();
-            void SetPageBreakAfter(string p);
-            string GetPageBreakAfter();
-            void SetCssText(string p);
-            string GetCssText();
-            void SetPixelTop(int p);
-            int GetPixelTop();
-            void SetPixelLeft(int p);
-            int GetPixelLeft();
-            void SetPixelWidth(int p);
-            int GetPixelWidth();
-            void SetPixelHeight(int p);
-            int GetPixelHeight();
-            void SetPosTop(float p);
-            float GetPosTop();
-            void SetPosLeft(float p);
-            float GetPosLeft();
-            void SetPosWidth(float p);
-            float GetPosWidth();
-            void SetPosHeight(float p);
-            float GetPosHeight();
-            void SetCursor(string p);
-            string GetCursor();
-            void SetClip(string p);
-            string GetClip();
-            void SetFilter(string p);
-            string GetFilter();
-            void SetAttribute(string strAttributeName, object AttributeValue, int lFlags);
-            object GetAttribute(string strAttributeName, int lFlags);
-            bool RemoveAttribute(string strAttributeName, int lFlags);
-        }
-
         [ComImport(),
          Guid("39088D7E-B71E-11D1-8F39-00C04FD946D0"),
          InterfaceType(ComInterfaceType.InterfaceIsIUnknown)
